% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_stats.R
\name{aaPosOmnibusTest}
\alias{aaPosOmnibusTest}
\title{Amino acid position omnibus test}
\usage{
aaPosOmnibusTest(object, aa_pos, correction = "bonferroni",
  n_correction = NULL)
}
\arguments{
\item{object}{An existing fit from a model function such as lm, glm and many
others.}

\item{aa_pos}{String specifying gene and amio acid position, example
\code{"A_9"}.}

\item{correction}{String specifying multiple testing correction method. See
details for further information.}

\item{n_correction}{Integer specifying number of comparisons to consider
during multiple testing correction calculations, must be at least equal to
number of comparisons being made; only set this (to non-default) when you
know what you are doing!}
}
\value{
Data frame containing omnibus test results for specified amino acid
  positions.
}
\description{
\code{aaPosOmnibusTest} calculate overall p-value for amino acid position
using likelihood rato test.
}
\details{
Likelihood ratio test is conducted by comparing model given in \code{object}
with extended model, that is created by including effect of amino acid
residues at each position in \code{aa_pos}. Amino acid effect is assumend to
be simply additive.
}
\examples{
hla_calls_file <- system.file("extdata", "HLAHD_output_example.txt", package = "MiDAS")
hla_calls <- readHlaCalls(hla_calls_file)
pheno_file <- system.file("extdata", "pheno_example.txt", package = "MiDAS")
pheno <- read.table(pheno_file, header = TRUE)
covar_file <- system.file("extdata", "covar_example.txt", package = "MiDAS")
covar <- read.table(covar_file, header = TRUE)
midas_data <- prepareMiDASData(hla_calls, pheno, covar, analysis_type = "aa_level")
object <- lm(OS ~ AGE + SEX, data = midas_data)
aaPosOmnibusTest(object, aa_pos = c("B_11", "E_107", "A_246"))

}
