% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{allele_frequencies}
\alias{allele_frequencies}
\title{Alleles frequencies scraped from allelefrequencies.net in wide format}
\format{A data frame with 2096 rows and 8 variables:
\describe{
  \item{allele}{allele number, character}
  \item{African American}{allele frequency in general reference population, float}
  \item{Amerindian}{allele frequency in general reference population, float}
  \item{Chinese}{allele frequency in general reference population, float}
  \item{European Caucasian}{allele frequency in general reference population, float}
  \item{Hispanic}{allele frequency in general reference population, float}
  \item{Indian}{allele frequency in general reference population, float}
  \item{Japanese}{allele frequency in general reference population, float}
}}
\source{
\url{www.allelefrequencies.net}
}
\usage{
allele_frequencies
}
\description{
A dataset containing allele frequencies across 5697 alleles
For quick yet more details I recommend the search results page in the
allelefrequencies.net database website. Data are gathered in wide format.
The general reference populations are defined based on different sources. For exampe
in case of HLA-A for Hispanic population frequencies from USA NMDP Hispanic South
or Central American are used, however in case of HLA-DQA1 it is Chile Santiago pop 2.
See \code{allele_frequencies_reference_populations}.
}
\keyword{datasets}
