% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{allele_frequencies_reference_populations}
\alias{allele_frequencies_reference_populations}
\title{Reference populations used in allele_frequencies}
\format{A data frame with 2096 rows and 8 variables:
\describe{
  \item{population}{general population name, character}
  \item{A}{names of the source populations used as reference for HLA gene A, character}
  \item{B}{names of the source populations used as reference for HLA gene B, character}
  \item{C}{names of the source populations used as reference for HLA gene C, character}
  \item{DQA1}{names of the source populations used as reference for HLA gene DQA1, character}
  \item{DQB1}{names of the source populations used as reference for HLA gene DQB1, character}
  \item{DPA1}{names of the source populations used as reference for HLA gene DPA1, character}
  \item{DPB1}{names of the source populations used as reference for HLA gene DPB1, character}
  \item{DRB1}{names of the source populations used as reference for HLA gene DRB1, character}
  \item{DRB3}{names of the source populations used as reference for HLA gene DRB3, character}
  \item{DRB4}{names of the source populations used as reference for HLA gene DRB4, character}
  \item{DRB5}{names of the source populations used as reference for HLA gene DRB5, character}
}}
\source{
\url{www.allelefrequencies.net}
}
\usage{
allele_frequencies_reference_populations
}
\description{
A dataset containing names of reference populations used for
extracting alleles frequency from allelefrequencies.net. For
some of the HLA genes different populations were used.
}
\keyword{datasets}
