% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_stats.R
\name{analyzeConditionalAssociations}
\alias{analyzeConditionalAssociations}
\title{Stepwise conditional association analysis}
\usage{
analyzeConditionalAssociations(object, variables, correction = "BH", th,
  keep = FALSE, rss_th = 1e-07, exponentiate = FALSE)
}
\arguments{
\item{object}{An existing fit from a model function such as lm, glm and many
others.}

\item{variables}{Character specifying variables to use in association tests.}

\item{correction}{String specifying multiple testing correction method. See
details for further information.}

\item{th}{number specifying p-value threshold for a variable to be considered
significant.}

\item{keep}{logical flag indicating if the output should be a list of results
resulting from each selection step. Default is to return only the final
result.}

\item{rss_th}{number specifying residual sum of squares threshold at which
function should stop adding additional variables. As the residual sum of
squares approaches \code{0} the perfect fit is obtained making further
attempts at variables selection nonsense, thus function is stopped. This
behavior can be controlled using \code{rss_th}.}

\item{exponentiate}{Logical indicating whether or not to exponentiate the
  coefficient estimates. Internally this is passed to \link[broom]{tidy}.
  This is typical for logistic and multinomial regressions, but a bad idea if
  there is no log or logit link. Defaults to FALSE.

\code{correction} specifies p-value adjustment method to use, common choice
is Benjamini & Hochberg (1995) (\code{"BH"}). Internally this is passed to
\link[stats]{p.adjust}. Check there to get more details.}
}
\value{
tibble with stepwise conditional testing results.
}
\description{
\code{analyzeConditionalAssociations} performs stepwise conditional testing
adding the previous top-associated variable as covariate, until there is no
more significant variables based on a self-defined threshold.
}
\details{
Selection criteria is the p-value from the test on coefficients values.
}
\examples{
library("survival")
hla_calls_file <- system.file("extdata", "HLAHD_output_example.txt", package = "MiDAS")
hla_calls <- readHlaCalls(hla_calls_file)
pheno_file <- system.file("extdata", "pheno_example.txt", package = "MiDAS")
pheno <- read.table(pheno_file, header = TRUE, stringsAsFactors = FALSE)
covar_file <- system.file("extdata", "covar_example.txt", package = "MiDAS")
covar <- read.table(covar_file, header = TRUE, stringsAsFactors = FALSE)
midas_data <- prepareHlaData(hla_calls = hla_calls,
                             pheno = pheno,
                             covar = covar,
                             inheritance_model = "additive"
)

## define base model with covariates only
object <- coxph(Surv(OS, OS_DIED) ~ AGE + SEX, data = midas_data)
analyzeConditionalAssociations(object,
                            variables = c("B*14:02", "DRB1*11:01"),
                            th = 0.05,
                            rss_th = 1e-07
)

}
