% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_stats.R
\name{analyzeHlaAssociations}
\alias{analyzeHlaAssociations}
\title{Association analysis of HLA alleles}
\usage{
analyzeHlaAssociations(model = "coxph", geno, pheno, covar,
  correction = "BH", output = "html", ...)
}
\description{
Association analysis of HLA alleles
}
\examples{
geno <- readHlaCalls("data/HLAexample.txt")
pheno <- read.table("data/pheno.txt",header=T)
covar <- read.table("data/covar.txt",header=T)
analyzeHlaAssociations(geno = geno, pheno = pheno, covar = covar)

}
