% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_stats.R
\name{analyzeMiDASData}
\alias{analyzeMiDASData}
\title{Analyze associations in MiDAS data}
\usage{
analyzeMiDASData(object, analysis_type = c("hla_allele", "aa_res",
  "aa_pos", "expression_level", "allele_g_group", "allele_supertype",
  "allele_group", "kir_genes", "hla_kir_interactions"), variables = NULL,
  conditional = FALSE, keep = FALSE, lower_frequency_cutoff = NULL,
  upper_frequency_cutoff = NULL, pvalue_cutoff = NULL,
  correction = "bonferroni", n_correction = NULL, logistic = NULL,
  binary_phenotype = NULL, th = 0.05, rss_th = 1e-07)
}
\arguments{
\item{object}{An existing fit from a model function such as lm, glm and many
others.}

\item{analysis_type}{String indicating the type of analysis being performed,
at this point it is only used for results formatting. Valid values are
\code{"hla_allele"}, \code{"aa_pos"}, \code{"aa_res"},
\code{"expression_level"}, \code{"allele_g_group"},
\code{"allele_supertype"}, \code{"allele_group"}, \code{"kir_genes"},
\code{"hla_kir_interactions"}.}

\item{variables}{Character specifying additional variables to use in
association tests except those choosen by \code{analysis_type}.}

\item{conditional}{Logical indicating if the analysis should be performed
using stepwise conditional tests or not. See
\link{analyzeConditionalAssociations} for more details.}

\item{keep}{logical flag indicating if the output should be a list of results
resulting from each selection step. Default is to return only the final
result.}

\item{lower_frequency_cutoff}{Number specifying lower threshold for inclusion
of a variable. If it's a number between 0 and 1 variables with frequency
below this number will not be considered during analysis. If it's greater
or equal 1 variables with number of counts less that this will not be
considered during analysis.}

\item{upper_frequency_cutoff}{Number specifying upper threshold for inclusion
of a variable. If it's a number between 0 and 1 variables with frequency
above this number will not be considered during analysis. If it's greater
or equal 1 variables with number of counts greater that this will not be
considered during analysis.}

\item{pvalue_cutoff}{Number specifying p-value cutoff for results to be
included in output. If \code{NULL} cutoff of \code{0.05} on
\code{p.adjusted} value is used instead.}

\item{correction}{String specifying multiple testing correction method. See
details for further information.}

\item{n_correction}{Integer specifying number of comparisons to consider
during multiple testing correction calculations, must be at least equql to
number of comparisons being made; only set this (to non-default) when you
know what you are doing!}

\item{logistic}{Logical indicating if statistical model used is logistic (eg.
\code{coxph}). If \code{NULL} function will try to figure this out. This is
only used for results formatting.}

\item{binary_phenotype}{Logical indicating if coefficient estimates should be
  exponentiated. This is typical for logistic and multinomial regressions,
  but a bad idea if there is no log or logit link. If \code{NULL} function
  will try to figure this out by testing if response is binary (\code{0} or
  \code{1}).

\code{variables} takes \code{NULL} as a default value, variables labeled with
specified \code{analysis_type} are used in association tests.

\code{correction} specifies p-value adjustment method to use, common choice
is Benjamini & Hochberg (1995) (\code{"BH"}). Internally this is passed to
\link[stats]{p.adjust}. Check there to get more details.}

\item{th}{number specifying p-value threshold for a variable to be considered
significant.}

\item{rss_th}{number specifying residual sum of squares threshold at which
function should stop adding additional variables. As the residual sum of
squares approaches \code{0} the perfect fit is obtained making further
attempts at variables selection nonsense, thus function is stopped. This
behavior can be controlled using \code{rss_th}.}
}
\value{
Tibble containing results for all tested variables.
}
\description{
\code{analyzeMiDASData} performs association analysis on MiDAS data using
statistical model specified by user. Function is intended for use with
\link{prepareMiDASData}. See examples section.
}
\examples{
library("survival")
hla_calls_file <- system.file("extdata", "HLAHD_output_example.txt", package = "MiDAS")
hla_calls <- readHlaCalls(hla_calls_file)
pheno_file <- system.file("extdata", "pheno_example.txt", package = "MiDAS")
pheno <- read.table(pheno_file, header = TRUE, stringsAsFactors = FALSE)
covar_file <- system.file("extdata", "covar_example.txt", package = "MiDAS")
covar <- read.table(covar_file, header = TRUE, stringsAsFactors = FALSE)
midas_data <- prepareMiDASData(hla_calls = hla_calls,
                               pheno = pheno,
                               covar = covar,
                               analysis_type = "hla_allele",
                               inheritance_model = "additive"
)

object <- coxph(Surv(OS, OS_DIED) ~ AGE + SEX, data = midas_data)
analyzeMiDASData(object, analysis_type = "hla_allele")

}
