% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_utils.R
\name{backquote}
\alias{backquote}
\title{Backquote character}
\usage{
backquote(x)
}
\arguments{
\item{x}{Character vector.}
}
\value{
Character vector with its elements backticked.
}
\description{
\code{backquote} places backticks around elements of character vector
}
\details{
\code{backquote} is useful when using HLA allele numbers in formulas, where
\code{'*'} and \code{':'} characters have special meanings.
}
