% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_n_pfiles.R
\name{call_n_pfiles}
\alias{call_n_pfiles}
\title{call raster Pisco files}
\usage{
call_n_pfiles(dir_pisco = "C:/Users/usuario 01/Desktop/Google drive",
  dekada_date = Sys.Date())
}
\arguments{
\item{dir_pisco}{a directory format}

\item{dekada_date}{a character object (a date in format "YYYY/MM/DD")}
}
\value{
a character object of several raster directories according to the date
}
\description{
calling raster Pisco files accordint to the date and directory
}
\examples{
call_n_pfiles(dekada_date = "2017-01-11")
call_n_pfiles(dekada_date = "2017-01-21")
call_n_pfiles(dekada_date = "2017-02-01")
}
\author{
Adrian Huerta
}
