% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_Perc_files.R
\name{call_Perc_files}
\alias{call_Perc_files}
\title{call raster percentile files}
\usage{

  call_Perc_files(dir_percent = "C:/Users/usuario 01/Desktop/Clim_dek_percentiles",
  dekada_date = "2017-02-21")
}
\arguments{
\item{dir_percent}{a directory format}

\item{dekada_date}{a character object (a date in format "YYYY/MM/DD")}
}
\value{
a list object of several raster directories according to the percentile values
}
\description{
calling raster percentile files accordint to the date and directory
}
\examples{
call_Perc_files(dekada_date = "2017-01-11")
call_Perc_files(dekada_date = "2017-01-21")
call_Perc_files(dekada_date = "2017-02-01")
}
\author{
Adrian Huerta
}
