% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_utils.R
\name{characterMatches}
\alias{characterMatches}
\title{Check if character matches one of possible values}
\usage{
characterMatches(x, choice)
}
\arguments{
\item{x}{character vector to test.}

\item{choice}{Character vector with possible values for \code{x}.}
}
\value{
Logical indicating if \code{x} matches one of the values in
  \code{choice}.
}
\description{
\code{characterMatches} checks if all elements of character matches values in
choices.
}
