% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_date_dek.R
\name{check_date_dek}
\alias{check_date_dek}
\title{checking date of calculation of dekedal amounts}
\usage{
check_date_dek(z)
}
\arguments{
\item{z}{a character object (a date in format "YYYY/MM/DD")}
}
\value{
a stop message if the date is different of 1, 11, or 21
}
\description{
checking date of calculation of dekedal amounts of precipitation
}
\examples{
check_date_dek("2017-01-01")
check_date_dek("2017-01-11")
}
\author{
Adrian Huerta
}
