% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_utils.R
\name{checkAdditionalData}
\alias{checkAdditionalData}
\title{Assert additional data}
\usage{
checkAdditionalData(data_frame, hla_calls, accept.null = FALSE)
}
\arguments{
\item{data_frame}{Data frame containing phenotype or covariate data
corresponding to accompanying hla calls data frame.}

\item{hla_calls}{Data frame containing HLA allele calls, as return by
\code{\link{readHlaCalls}} function.}

\item{accept.null}{Logical indicating if NULL data_frame should be accepted.}
}
\value{
Logical indicating if \code{data_frame} is properly formatted.
  Otherwise raise error.
}
\description{
\code{checkAdditionalData} asserts if phenotype or covariate data frame
has proper format.
}
\examples{
hla_file <- system.file("extdata", "HLAHD_output_example.txt", package = "MiDAS")
pheno_file <- system.file("extdata", "pheno_example.txt", package = "MiDAS")
pheno <- read.table(pheno_file, header = TRUE)
hla_calls <- readHlaCalls(hla_file)
checkAdditionalData(pheno, hla_calls)

}
