% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_asserts.R
\name{checkColDataFormat}
\alias{checkColDataFormat}
\title{Assert colData data}
\usage{
checkColDataFormat(data_frame)
}
\arguments{
\item{data_frame}{Data frame containing colData data used to construct
\code{\link{MiDAS}} object.}
}
\value{
Logical indicating if \code{data_frame} is properly formatted.
  Otherwise raise an error.
}
\description{
\code{checkColDataFormat} asserts if the colData data frame has proper format.
}
\seealso{
Other assert functions: \code{\link{hasTidyMethod}},
  \code{\link{isClassOrNULL}}, \code{\link{isCountOrNULL}},
  \code{\link{isExperimentCountsOrZeros}},
  \code{\link{isTRUEorFALSE}},
  \code{\link{objectHasPlaceholder}}
}
\concept{assert functions}
