% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_utils.R
\name{checkHlaCallsFormat}
\alias{checkHlaCallsFormat}
\title{Assert hla calls data frame format}
\usage{
checkHlaCallsFormat(hla_calls)
}
\arguments{
\item{hla_calls}{Data frame containing HLA allele calls, as return by
\code{\link{readHlaCalls}} function.}
}
\value{
Logical indicating if \code{hla_calls} follows hla calls data frame
  format. Otherwise raise error.
}
\description{
\code{checkHlaCallsFormat} asserts if hla calls data frame have proper
format.
}
\examples{
file <- system.file("extdata", "HLAHD_output_example.txt", package = "MiDAS")
hla_calls <- readHlaCalls(file)
checkHlaCallsFormat(hla_calls)

}
