% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_utils.R
\name{checkKirCallsFormat}
\alias{checkKirCallsFormat}
\title{Assert KIR counts data frame format}
\usage{
checkKirCallsFormat(kir_calls, accept.null = FALSE)
}
\arguments{
\item{kir_calls}{Data frame containing KIR gene counts, as returned by
\code{\link{readKIRCalls}} function.}

\item{accept.null}{Logical indicating if NULL \code{kir_calls} should be
accepted.}
}
\value{
Logical indicating if \code{kir_calls} follow KIR counts data frame
  format. Otherwise raise error.
}
\description{
\code{checkKirCallsFormat} asserts if KIR counts data frame have proper
format.
}
\examples{
file <- system.file("extdata", "KPI_output_example.txt", package = "MiDAS")
kir_calls <- readKIRCalls(file)
checkKirCallsFormat(kir_calls)

}
\seealso{
\code{\link{readKIRCalls}}, \code{\link{getHlaKirInteractions}},
  \code{\link{kirHaplotypeToCounts}}, \code{\link{prepareMiDAS}}.

Other assert functions: \code{\link{characterMatches}},
  \code{\link{checkAdditionalData}},
  \code{\link{checkColDataFormat}},
  \code{\link{checkHlaCallsFormat}},
  \code{\link{checkStatisticalModel}},
  \code{\link{colnamesMatches}},
  \code{\link{functionExists}},
  \code{\link{hasTidyMethod}},
  \code{\link{isCharacterOrNULL}},
  \code{\link{isClassOrNULL}}, \code{\link{isClass}},
  \code{\link{isCountOrNULL}},
  \code{\link{isCountsOrZeros}},
  \code{\link{isExperimentCountsOrZeros}},
  \code{\link{isFlagOrNULL}},
  \code{\link{isInheritanceModelApplicable}},
  \code{\link{isNumberOrNULL}},
  \code{\link{isStringOrNULL}},
  \code{\link{isTRUEorFALSE}},
  \code{\link{objectHasPlaceholder}},
  \code{\link{stringMatches}},
  \code{\link{validateFrequencyCutoffs}}
}
\concept{assert functions}
