% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_utils.R
\name{checkKirCountsFormat}
\alias{checkKirCountsFormat}
\title{Assert kir counts data frame format}
\usage{
checkKirCountsFormat(kir_counts, accept.null = FALSE)
}
\arguments{
\item{kir_counts}{Data frame containing KIR genes counts, as return by
\code{\link{readKirCalls}} function.}

\item{accept.null}{Logical indicating if NULL \code{kir_counts} should be
accepted.}
}
\value{
Logical indicating if \code{kir_counts} follows kir counts data frame
  format. Otherwise raise error.
}
\description{
\code{checkKirCountsFormat} asserts if kir counts data frame have proper
format.
}
\examples{
file <- system.file("extdata", "KIP_output_example.txt", package = "MiDAS")
kir_counts <- readKirCalls(file)
checkKirCountsFormat(kir_counts)

}
