% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_utils.R
\name{checkKirGenesFormat}
\alias{checkKirGenesFormat}
\title{Check KIR genes format}
\usage{
checkKirGenesFormat(genes)
}
\arguments{
\item{genes}{Character vector containing KIR gene names.}
}
\value{
Logical vector specifying if values of \code{genes} follows KIR genes
  naming conventions.
}
\description{
\code{checkKirGenesFormat} test if the input character contains correct KIR
gene names
}
\examples{
genes <- c("KIR3DL3", "KIR2DS2", "KIR2DL2")
checkKirGenesFormat(genes)

}
