% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_utils.R
\name{checkStatisticalModel}
\alias{checkStatisticalModel}
\title{Assert statistical model}
\usage{
checkStatisticalModel(object)
}
\arguments{
\item{object}{An existing fit from a model function such as lm, glm and many
others.}
}
\value{
Logical indicating if \code{data_frame} is an existing fit from a
model function such as lm, glm and many others. Otherwise raise error.
}
\description{
\code{checkStatisticalModel} asserts if object is an existing fit from a
model function such as lm, glm and many others.
}
\examples{
object <- lm(dist ~ speed, data = cars)
checkStatisticalModel(object)

}
