% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_utils.R
\name{colnamesMatches}
\alias{colnamesMatches}
\title{Check column names}
\usage{
colnamesMatches(x, cols)
}
\arguments{
\item{x}{Data frame vector to test.}

\item{cols}{Ordered character vector with values for \code{x} colnames
to test.}
}
\value{
Logical indicating if \code{x} colnames matches values in
  \code{choice}.
}
\description{
\code{colnamesMatches} checks columns of data frame are named as specified
}
