% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{correct_pool}
\alias{correct_pool}
\title{Correct Pool}
\usage{
correct_pool(pool_data, npcs = 3)
}
\arguments{
\item{pool_data}{tibble containing protein, metabolite, and log2_abundance}

\item{number}{of principal components to subtract}
}
\value{
a list containing:
\describe{
 \item{pool_log2_matrix}{A matrix of uncorrected abundances}
 \item{scree_data}{Percent variance explained for each PC}
 \item{top_pc_fit}{Projection onto top principal components}
 \item{log2_pc_projection}{\code{pool_log2_matrix} - \code{top_pc_fit}}
 \item{log2_pc_projection_tbl}{tibble of log2_pc_projection}
}
}
\description{
Subtracts the top principal components from a matrix of proteins x metabolites
}
