% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_transformationFunctions.R
\name{countsToHlaCalls}
\alias{countsToHlaCalls}
\title{Convert HLA counts table to HLA calls}
\usage{
countsToHlaCalls(counts)
}
\arguments{
\item{counts}{Data frame with HLA alleles counts, as returned by
\code{\link{hlaCallsToCounts}} function. First column should contain
samples IDs, following columns should be named with valid HLA alleles
numbers.}
}
\value{
Data frame containing HLA allele calls.
}
\description{
\code{countsToHlaCalls} convert counts table to HLA calls data frame, this
is useful when working with data from UK Biobank.
}
\details{
Note that proper HLA calls reconstruction from counts table is only possible
under additive inheritance model. This mode of operation is the only one
implemented so the function will always treat counts table as coming from
\code{hlaCallsToCounts(hla_calls, inheritance_model = 'additive')}.
}
\examples{
file <- system.file("extdata", "HLAHD_output_example.txt", package = "MiDAS")
hla_calls <- readHlaCalls(file)
hla_counts <- hlaCallsToCounts(hla_calls, inheritance_model = "additive")
countsToHlaCalls(hla_counts)

}
