% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dekadal_cal.R
\name{dekadal_cal}
\alias{dekadal_cal}
\title{calculation of dekadal amount}
\usage{
dekadal_cal(r_files = call_n_pfiles())
}
\arguments{
\item{r_files}{a vector several raster directories (output of \link[MIDAS]{call_n_pfiles})}
}
\value{
a raster file object
}
\description{
calculation of dekadal amount of precipitation
}
\author{
Adrian Huerta
}
