% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dict_dist_grantham}
\alias{dict_dist_grantham}
\title{Grantham distance}
\format{
Named integer vector of length 400.
}
\usage{
dict_dist_grantham
}
\description{
Integer vector giving Grantham distance values between pairs of amino acid
residues.
}
\keyword{datasets}
