% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_utils.R
\name{distGrantham}
\alias{distGrantham}
\title{Grantham distance}
\usage{
distGrantham(aa1, aa2)
}
\arguments{
\item{aa1}{Character vector giving amino acid sequence using one letter
codings. Each element must correspond to single amino acid.}

\item{aa2}{Character vector giving amino acid sequence using one letter
codings. Each element must correspond to single amino acid.}
}
\value{
Integer normalized Grantham distance between \code{aa1} and \
  code{aa2}.
}
\description{
\code{distGrantham} calculates normalized Grantham distance between two
amino acid sequences.
}
\details{
Distance between amino acid sequences is normalized by length of compared
sequences.

Lengths of \code{aa1} and \code{aa2} must be equal.
}
\examples{
distGrantham(
  aa1 = c("A", "S", "W"),
  aa2 = c("A", "S", "V")
)

}
