% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drought_class.R
\name{drought_class}
\alias{drought_class}
\title{drougth clasification function}
\usage{
drought_class(r_dek, p10, p20, p30, p70, p80, p90)
}
\arguments{
\item{r_dek}{a raster object (a dekadal precipitation raster)}

\item{p10}{a raster object (percentile 10th of climatology dekadal precipitation)}

\item{p20}{a raster object (percentile 20th of climatology dekadal precipitation)}

\item{p30}{a raster object (percentile 30th of climatology dekadal precipitation)}

\item{p70}{a raster object (percentile 70th of climatology dekadal precipitation)}

\item{p80}{a raster object (percentile 80th of climatology dekadal precipitation)}

\item{p90}{a raster object (percentile 90th of climatology dekadal precipitation)}
}
\value{
a raster object
}
\description{
drougth clasification function for dekadal precipitation
}
\seealso{
\url{http://www.crc-sas.org/es/content/monitoreo/reporte_sequias.pdf}
}
\author{
Adrian Huerta
}
