% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_class.R
\name{experimentMatToDf}
\alias{experimentMatToDf}
\title{Helper transform experimentt matrix to data frame}
\usage{
experimentMatToDf(mat)
}
\arguments{
\item{mat}{Matrix}
}
\value{
Data frame
}
\description{
Function transpose \code{mat} and inserts columnames of input \code{mat} as
a 'ID' column.
}
