% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_transformationFunctions.R
\name{filterExperimentByFrequency}
\alias{filterExperimentByFrequency}
\title{Filter experiment by frequency}
\usage{
filterExperimentByFrequency(experiment, carrier_frequency = FALSE,
  lower_frequency_cutoff = NULL, upper_frequency_cutoff = NULL)
}
\arguments{
\item{experiment}{Matrix}

\item{lower_frequency_cutoff}{Number of positive value or \code{NULL}}

\item{upper_frequency_cutoff}{Number of positive value or \code{NULL}}
}
\value{
Matrix
}
\description{
Helper function for experiments filtering
}
