% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_transformationFunctions.R
\name{filterExperimentByVariables}
\alias{filterExperimentByVariables}
\title{Filter experiment by frequency}
\usage{
filterExperimentByVariables(experiment, variables)
}
\arguments{
\item{experiment}{matrix}

\item{variables}{character}
}
\value{
Matrix
}
\description{
Helper function for experiments filtering
}
