% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_class.R
\name{filterMiDAS}
\alias{filterMiDAS}
\title{Filter midas object}
\usage{
filterMiDAS(object, filter_by = c("hla_allele_frequency"), ...)
}
\arguments{
\item{filter}{Character giving suffix of filter function.}
}
\value{
Object of class MiDAS.
}
\description{
Filter midas object
}
