% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_transformationFunctions.R
\name{formatResults}
\alias{formatResults}
\title{Helper function for pretty formating statistical analysis results}
\usage{
formatResults(results, filter_by = "p.value <= 0.05",
  arrange_by = "p.value", select_cols = c("term", "estimate",
  "std.error", "p.value", "p.adjusted"), format = c("html", "latex"),
  header = NULL)
}
\arguments{
\item{results}{Tibble as returned by \code{\link{analyzeAssociations}}.}

\item{filter_by}{Character vector specifying conditional expression used to
filter \code{results}, this is equivalent to \code{...} argument passed to
\code{\link[dplyr]{filter}} except it has to be a character vector.}

\item{arrange_by}{Character vector specifying variable names to use for
sorting. Equivalent to \code{...} argument passed to
\code{\link[dplyr]{arrange}}.}

\item{select_cols}{Character vector specifying variable names that should be
included in the output table. Can be also used to rename selected
variables, see examples.}

\item{format}{String with possible values \code{"latex"} and \code{"html"}.}

\item{header}{String specifying header for result table. If \code{NULL}
no header is added.}
}
\value{
Character vector of formatted table source code.
}
\description{
\code{formatResults} format statistical analysis results table to html or
latex format.
}
\examples{
hla_calls <- readHlaCalls(system.file("extdata", "HLAHD_output_example.txt", package = "MiDAS"))
pheno <- read.table(
  system.file("extdata", "pheno_example.txt", package = "MiDAS"),
  header = TRUE)
midas_data <- prepareMiDAS(hla_calls, pheno, analysis_type = "hla_allele")
object <- lm(OS ~ 1 + term, data = midas_data)
res <- analyzeAssociations(object, variables = colnames(midas_data)[-1])
formatResults(res,
              filter_by = c("p.value <= 0.05", "estimate > 0"),
              arrange_by = c("p.value * estimate"),
              select_cols = c("allele" = "term", "p.value"),
              format = "html",
              header = "HLA allelic associations")

}
\seealso{
\code{\link{runMiDAS}}, \code{\link{analyzeAssociations}},
  \code{\link{analyzeConditionalAssociations}}.
}
