% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_stats.R
\name{forwardAllelesSelection}
\alias{forwardAllelesSelection}
\title{Stepwise forward alleles subset selection}
\usage{
forwardAllelesSelection(object, scope, th, test = c("F", "Chisq"),
  rss_th = 1e-07, ...)
}
\arguments{
\item{object}{object fitted by some model-fitting function.}

\item{scope}{formula specifying a maximal model which should include the
current one. All additional terms in the maximal model with all marginal
terms in the original model are tried.}

\item{th}{number specifying p-value threshold for a term to be included into
model.}

\item{test}{string indicating test statistic to use for p-value calculation.
Can be either "F" or "Chisq".}

\item{rss_th}{number specifying residual sum of squares threshold at which
function should stop adding additional terms.}

\item{...}{additional parameters passsed to \link[MASS]{addterm} function.

All the variables in the \code{scope} should be defined in the \code{object}.

The F test is only appropriate for lm and aov models, and perhaps for some
over-dispersed glm models. The Chisq test can be an exact test (lm models
with known scale) or a likelihood-ratio test depending on the method.

As the residual sum of squares approaches \code{0} the perfect fit is
obtained making further attempts at model selection nonsense, thus function
is stopped. This behavior can be controlled using \code{rss_th}.}
}
\value{
selected model of the same class as \code{object} is returned.
}
\description{
\code{forwardAllelesSelection} does stepwise conditional testing adding the
previous top-associated allele as covariate, until there’s no more
significant alleles using a self-defined threshold.
}
\examples{
hla_calls_file <- system.file("extdata", "HLAHD_output_example.txt", package = "MiDAS")
hla_calls <- readHlaCalls(hla_calls_file)
pheno_file <- system.file("extdata", "pheno_example.txt", package = "MiDAS")
pheno <- read.table(pheno_file, header = TRUE)
covar_file <- system.file("extdata", "covar_example.txt", package = "MiDAS")
covar <- read.table(covar_file, header = TRUE)
hla_data <- prepareHlaData(hla_calls, pheno, covar)
object <- survival::coxph(survival::Surv(OS, OS_DIED) ~ AGE + SEX, data = hla_data$data)
scope <- survival::Surv(OS, OS_DIED) ~ AGE + SEX + `B*57:01` + `C*07:02`
forwardAllelesSelection(object, scope, th = 0.05, test = "Chisq")

}
