% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_stats.R
\name{forwardConditionalSelection}
\alias{forwardConditionalSelection}
\title{Stepwise conditional variables selection}
\usage{
forwardConditionalSelection(object, variables, th, keep = FALSE,
  rss_th = 1e-07)
}
\arguments{
\item{object}{An existing fit from a model function such as lm, glm and many
others.}

\item{variables}{Character specifying variables to use in selection
procedure.}

\item{th}{number specifying p-value threshold for a term to be included into
model.}

\item{keep}{logical flag indicating if the output should be a list of models
resulting from each selection step. Default is to return only the final
model.}

\item{rss_th}{number specifying residual sum of squares threshold at which
  function should stop adding additional terms.

As the residual sum of squares approaches \code{0} the perfect fit is
obtained making further attempts at model selection nonsense, thus function
is stopped. This behavior can be controlled using \code{rss_th}.}
}
\value{
selected model or list of models. See \code{keep} parameter.
}
\description{
\code{stepwiseConditionalSelection} performs stepwise conditional testing
adding the previous top-associated variable as covariate, until there’s no
more significant variables based on a self-defined threshold.
}
\details{
Selection criteria is the p-value from the test on coefficients values.
}
\examples{
library("survival")
hla_calls_file <- system.file("extdata", "HLAHD_output_example.txt", package = "MiDAS")
hla_calls <- readHlaCalls(hla_calls_file)
pheno_file <- system.file("extdata", "pheno_example.txt", package = "MiDAS")
pheno <- read.table(pheno_file, header = TRUE)
covar_file <- system.file("extdata", "covar_example.txt", package = "MiDAS")
covar <- read.table(covar_file, header = TRUE)
midas_data <- prepareHlaData(hla_calls = hla_calls,
                             pheno = pheno,
                             covar = covar,
                             inheritance_model = "additive"
)

## define base model with covariates only
object <- coxph(Surv(OS, OS_DIED) ~ AGE + SEX, data = midas_data)
forwardConditionalSelection(object,
                            variables = c("B*14:02", "DRB1*11:01"),
                            th = 0.05,
                            keep = FALSE,
                            rss_th = 1e-07
)

}
