% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{generate_met_z_scores}
\alias{generate_met_z_scores}
\title{Generate Metabolite Z Scores}
\usage{
generate_met_z_scores(met_data, sd_method = "holdout", quantile_range = 0.5)
}
\arguments{
\item{met_data}{a tibble containing metabolite, protein and log2_abundance_corrected}

\item{sd_method}{method to use when estimating standard deviation:
\describe{
 \item{holdout}{For a given value, calculate SD by removing that value and calculating SD over all other measurements of that metabolite. Conservative when there are many enrichments or depletions of a given metabolite.}
 \item{central_quantiles}{Estimate SD from the central quantiles of the
 relative abundance distribution. For intuition, 66 \% of Normally-distribute
 observations are expected to fall within 1 SD of the mean. So, the range of
 the central 66 \% of the Normal distribution would approximate twice the
 standard deviation.}
 }}

\item{quantile_range}{For \code{central_quantiles}, fraction of relative abundance distribution to use when estimating SD.}
}
\value{
met_data containing estimates of the SD of an observation and p-values.
}
\description{
Estimate the mean, and SD of abundances changes and resulting statistical significance based on a Wald test.
}
