% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_transformationFunctions.R
\name{getAAFrequencies}
\alias{getAAFrequencies}
\title{Calculate amino acid's frequencies}
\usage{
getAAFrequencies(aa_variation)
}
\arguments{
\item{aa_variation}{Data frame holding amino acid variation data as returned
by \link{hlaToAAVariation}.}
}
\value{
Data frame containing the amino acid's positions and their
  corresponding frequencies.
}
\description{
\code{getAAFrequencies} calculates amino acid's frequencies in amino acid
variations data frame.
}
\details{
Amino acid's frequencies are counted in reference to sample taking both gene
copies into consideration. `n / (2 * j)` where `n` is the number of amino
acid occurrences and `j` is the sample size.
}
\examples{
file <- system.file("extdata", "HLAHD_output_example.txt", package = "MiDAS")
hla_calls <- readHlaCalls(file)
aa_variation <- hlaToAAVariation(hla_calls)
getAAFrequencies(aa_variation)

}
\seealso{
\code{\link{hlaToAAVariation}}
}
