% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_utils.R
\name{getAlleleResolution}
\alias{getAlleleResolution}
\title{Infers HLA allele resolution}
\usage{
getAlleleResolution(allele, na.pass = TRUE)
}
\arguments{
\item{allele}{Character vector containing HLA allele numbers.}

\item{na.pass}{Logical indicating if \code{NA} values should be accepted. If
set to \code{FALSE} function will rise an error when
\code{allele} will contain \code{NA} values.}
}
\value{
Integer vector specifying alleles resolutions.
}
\description{
Infers HLA allele resolution
}
\examples{
allele <- c("A*01:01", "A*01:02")
getAlleleResolution(allele)

}
