% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_transformationFunctions.R
\name{getCountsFrequencies}
\alias{getCountsFrequencies}
\title{Calculate variables frequencies}
\usage{
getCountsFrequencies(counts_table, inheritance_model = c("dominant",
  "recessive", "additive"))
}
\arguments{
\item{counts_table}{Data frame containing variables counts, such as produced
by \code{\link{hlaCallsToCounts}}.}

\item{inheritance_model}{String specifying inheritance model to use.
Available choices are \code{"dominant"}, \code{"recessive"},
\code{"additive"}. In \code{"dominant"} model homozygotes and heterozygotes
are coded as \code{1}. In \code{"recessive"} model homozygotes are coded as
\code{1} and all other as \code{0}. In \code{"additive"} model homozygotes
are coded as \code{2} and heterozygotes as \code{1}.}
}
\value{
Data frame containing variables, its corresponding total counts
  and frequencies.
}
\description{
\code{getCountsFrequencies} calculate variables frequencies based on counts
table, such as produced by \code{\link{hlaCallsToCounts}}.
}
\details{
Variables frequencies are counted in reference to sample size, depending on
the inheritance model under which the counts table has been generated one
might need to take under consideration both gene copies. Here sample size is
assumed to be depended on both gene copies for \code{"additive"} inheritance
model (`n / (2 * j)` where `n` is the number of term occurrences and `j`
is the sample size). For other models the sample size is taken as is
(`n / j`).
}
\examples{
\dontrun{
file <- system.file("extdata", "HLAHD_output_example.txt", package = "MiDAS")
hla_calls <- readHlaCalls(file)
hla_counts <- hlaCallsToCounts(hla_calls, inheritance_model = "additive")
getCountsFrequencies(hla_counts)
}

}
\seealso{
\code{\link{hlaCallsToCounts}}
}
