% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_class.R
\docType{methods}
\name{getExperiments}
\alias{getExperiments}
\alias{getExperiments,MiDAS-method}
\title{Extract available analysis types from MiDAS object.}
\usage{
getExperiments(object)

\S4method{getExperiments}{MiDAS}(object)
}
\arguments{
\item{object}{\code{\link{MiDAS}} object}
}
\value{
Character giving names of available analysis types.
}
\description{
Extract available analysis types from MiDAS object.
}
