% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_transformationFunctions.R
\name{getFrequencyMask}
\alias{getFrequencyMask}
\title{Helper function}
\usage{
getFrequencyMask(df, lower_frequency_cutoff = NULL,
  upper_frequency_cutoff = NULL)
}
\description{
Helper function
}
