% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_utils.R
\name{getHlaCallsGenes}
\alias{getHlaCallsGenes}
\title{Get HLA calls genes}
\usage{
getHlaCallsGenes(hla_calls)
}
\arguments{
\item{hla_calls}{HLA calls data frame, as returned by
\code{\link{readHlaCalls}} function.}
}
\value{
Character vector of genes in \code{hla_calls}.
}
\description{
\code{getHlaCallsGenes} get's genes found in HLA calls.
}
