% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_class.R
\docType{methods}
\name{getKirCalls}
\alias{getKirCalls}
\alias{getKirCalls,MiDAS-method}
\title{Extract KIR calls from MiDAS  object.}
\usage{
getKirCalls(object)

\S4method{getKirCalls}{MiDAS}(object)
}
\arguments{
\item{a}{}
}
\value{
String giving name of object's inheritance model.
}
\description{
Extract KIR calls from MiDAS  object.
}
