% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_transformationFunctions.R
\name{getKIRFrequencies}
\alias{getKIRFrequencies}
\title{Calculate KIR genes frequencies}
\usage{
getKIRFrequencies(kir_calls)
}
\arguments{
\item{kir_calls}{KIR calls data frame, as returned by
\code{\link{readKIRCalls}} function.}
}
\value{
Data frame containing KIR genes and their corresponding frequencies.
}
\description{
\code{getKIRFrequencies} calculates KIR genes frequencies in KIR calls data
frame.
}
\examples{
getKIRFrequencies(MiDAS_tut_KIR)

}
