% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_utils.R
\name{getObjectDetails}
\alias{getObjectDetails}
\title{Get attributes of statistical model object}
\usage{
getObjectDetails(object)
}
\arguments{
\item{object}{An existing fit from a model function such as lm, glm and many
others.}
}
\value{
List with following elemnts:
\describe{
  \item{formula_vars}{Character containing names of variables in object
    formula}
  \item{data}{MiDAS object associated with model}
  \item{data_vars}{Character containing names of variables in object data}
}
}
\description{
\code{getObjectDetails} extracts some of the statistical model object
attributes that are needed for \code{runMiDAS} internal calculations.
}
\details{
TODO write unit tests
}
