% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_class.R
\docType{methods}
\name{getPlaceholder}
\alias{getPlaceholder}
\alias{getPlaceholder,MiDAS-method}
\title{Extract placeholder name from MiDAS object.}
\usage{
getPlaceholder(object)

\S4method{getPlaceholder}{MiDAS}(object)
}
\arguments{
\item{object}{\code{\link{MiDAS}} object.}
}
\value{
String giving object's placeholder.
}
\description{
Extract placeholder name from MiDAS object.
}
