% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_utils.R
\name{getReferenceFrequencies}
\alias{getReferenceFrequencies}
\title{Get HLA allele frequencies in reference populations}
\usage{
getReferenceFrequencies(ref, pop, carrier_frequency = FALSE)
}
\arguments{
\item{ref}{Data frame}

\item{pop}{Character}

\item{carrier_frequency}{Logical}
}
\value{
Data frame
}
\description{
Get HLA allele frequencies in reference populations
}
