% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_utils.R
\name{hasTidyMethod}
\alias{hasTidyMethod}
\title{Check if tidy method for class exist}
\usage{
hasTidyMethod(class)
}
\arguments{
\item{class}{Object class.}
}
\value{
Logical indicating if there is tidy method for given class.
}
\description{
\code{hasTidyMethod} checks if there is tidy method for given class.
}
