% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{heatmap_plot}
\alias{heatmap_plot}
\title{Heatmap Plot}
\usage{
heatmap_plot(dat)
}
\arguments{
\item{dat}{A matrix of abuundances}
}
\value{
a grob
}
\description{
Creates a heatmap with hierarchically clusters rows and columns.
}
