% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_utils.R
\name{hlaAlignmentGrantham}
\alias{hlaAlignmentGrantham}
\title{Helper function alignment for Grantham distance calculations}
\usage{
hlaAlignmentGrantham(gene, resolution)
}
\arguments{
\item{alignment}{Matrix containing HLA alleles amino acids level alignment.}
}
\description{
\code{hlaAlignmentGrantham} get HLA alignment processed so that grantham
distance between alleles can be calculated. Processing includes extracting
exons 1 and 2, masking indels, gaps and stop codons.
}
