% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_stats.R
\name{hlaAssocModel}
\alias{hlaAssocModel}
\title{Association models for analysis of HLA alleles}
\usage{
hlaAssocModel(model, response, variable, data, ...)
}
\arguments{
\item{model}{String specifying statistical model to use.}

\item{response}{String specifying response variable in \code{data}.}

\item{variable}{Character specifying variables in \code{data}.}

\item{data}{Data frame containing the variables in the model.}

\item{...}{Further arguments passed to \code{model} function.}
}
\value{
Fit from specified \code{model} function.
}
\description{
\code{hlaAssocModels} make creating statistical models for HLA association
analyses easier.
}
\examples{
library("survival")

hla_calls_file <- system.file("extdata", "HLAHD_output_example.txt", package = "MiDAS")
hla_calls <- readHlaCalls(hla_calls_file)
pheno_file <- system.file("extdata", "pheno_example.txt", package = "MiDAS")
pheno <- read.table(pheno_file, header = TRUE)
covar_file <- system.file("extdata", "covar_example.txt", package = "MiDAS")
covar <- read.table(covar_file, header = TRUE)
hla_data <- prepareHlaData(hla_calls, pheno, covar, inheritance_model = "additive")
hlaAssocModel(model = "coxph",
              response = "Surv(OS, OS_DIED)",
              variable = c("AGE", "SEX"),
              data = hla_data
)

}
