% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_transformationFunctions.R
\name{hlaToVariable}
\alias{hlaToVariable}
\title{Converts hla calls data frame according to match table}
\usage{
hlaToVariable(hla_calls, dictionary, nacols.rm = TRUE)
}
\arguments{
\item{hla_calls}{data frame containing HLA allele calls, in a format as
return by `readHlaCalls` function.}

\item{dictionary}{table containing allele numbers matchings, this can be
either path to tsv file or a data frame. Fist column
should contain alleles and second corresponding
assignments. Optionally matchings files shipped with the
package can be referred to by using one of following
strings: "2digit_A-allele_expression",
"2digit_C-allele_expression", "4digit_allele_Ggroup",
"4digit_B-allele_Bw", "4digit_C-allele_C1-2",
"4digit_supertype".}

\item{nacol.rm}{logical indicating if result columns that contain only
\code{NA} should be removed.}
}
\value{
Data frame of hla numbers converted to additional variables according
        to match table.
}
\description{
Converts hla calls data frame according to match table
}
\examples{
hla_calls <- system.file("extdata/HLAHD_output_example.txt", package = "MiDAS")
hla_calls <- readHlaCalls(hla_calls)
hlaToVariable(hla_calls, dictionary = "4digit_supertype")

}
