% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midas_touch.R
\name{html_to_list}
\alias{html_to_list}
\title{Uses xml2 to parse html text into a list}
\usage{
html_to_list(x)
}
\arguments{
\item{x}{string of html}
}
\value{
html as an R list
}
\description{
Uses xml2 to parse html text into a list
}
\examples{
# create some test data
html <- '<div class="example"><h3>test</h3></div>'
html_to_list(html)
}

