% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_stats.R
\name{HWETest}
\alias{HWETest}
\title{Test for Hardy Weinberg Equilibrium}
\usage{
HWETest(object, experiment = c("hla_alleles", "hla_aa", "hla_g_groups",
  "hla_supertypes", "hla_NK_ligands"), HWE_group = NULL,
  HWE_cutoff = NULL, as.MiDAS = FALSE)
}
\arguments{
\item{X}{Matrix with genotype counts, with samples as columns and alleles as
rows.}
}
\value{
A vector of p-values
}
\description{
Test HLA calls experiment (matrix) for Hardy Weinberg Equilibrium. This
function \code{HWChisqStats} from \code{\link{HardyWeinberg}} package.
}
