% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_utils.R
\name{isCharacterOrNULL}
\alias{isCharacterOrNULL}
\title{Check if object is character vector or NULL}
\usage{
isCharacterOrNULL(x)
}
\arguments{
\item{x}{object to test.}
}
\value{
Logical indicating if object is character vector or NULL
}
\description{
\code{isCharacterOrNULL} checks if object is character vector or NULL.
}
