% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_utils.R
\name{isClassOrNULL}
\alias{isClassOrNULL}
\title{Check if object is of class x or null}
\usage{
isClassOrNULL(x, class)
}
\arguments{
\item{x}{object to test.}

\item{class}{String specifying class to test.}
}
\value{
Logical indicating if \code{x} is an instance of \code{class}.
}
\description{
\code{isClassOrNULL} checks if object is an instance of a specified class or
is null.
}
\seealso{
Other assert functions: \code{\link{characterMatches}},
  \code{\link{checkAdditionalData}},
  \code{\link{checkHlaCallsFormat}},
  \code{\link{checkKirCountsFormat}},
  \code{\link{checkStatisticalModel}},
  \code{\link{colnamesMatches}},
  \code{\link{hasTidyMethod}},
  \code{\link{isCharacterOrNULL}},
  \code{\link{isCountOrNULL}},
  \code{\link{isCountsOrZeros}},
  \code{\link{isFlagOrNULL}}, \code{\link{isNumberOrNULL}},
  \code{\link{isStringOrNULL}},
  \code{\link{isTRUEorFALSE}}, \code{\link{stringMatches}}
}
\concept{assert functions}
