% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_utils.R
\name{isCountsOrZeros}
\alias{isCountsOrZeros}
\title{Check if vectors contains only counts or zeros}
\usage{
isCountsOrZeros(x, na.rm = TRUE)
}
\arguments{
\item{x}{Numeric vector or object that can be \code{unlist} to numeric
vector.}

\item{na.rm}{Logical indicating if \code{NA} values should be omited.}
}
\value{
Logical indicating if provided vector contains only positive integers
  or zeros.
}
\description{
\code{isCountsOrZeros} checks if vector contains only positive integers or
zeros.
}
\examples{
isCountsOrZeros(c(0, 1, 2))

}
