% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_utils.R
\name{isExperimentCountsOrZeros}
\alias{isExperimentCountsOrZeros}
\title{Check if frequencies can be calculated for an experiment}
\usage{
isExperimentCountsOrZeros(x, na.rm = TRUE)
}
\arguments{
\item{x}{Matrix or SummarizedExperiment object.}

\item{na.rm}{Logical indicating if \code{NA} values should be accepted.}
}
\value{
Logical indicating if \code{x} contains only positive integers or
  zeros.
}
\description{
\code{isExperimentCountsOrZeros} checks if experiment contains only positive
integers or zeros.
}
\seealso{
Other assert functions: \code{\link{characterMatches}},
  \code{\link{checkAdditionalData}},
  \code{\link{checkColDataFormat}},
  \code{\link{checkHlaCallsFormat}},
  \code{\link{checkKirCallsFormat}},
  \code{\link{checkKirCountsFormat}},
  \code{\link{checkStatisticalModel}},
  \code{\link{colnamesMatches}},
  \code{\link{functionExists}},
  \code{\link{hasTidyMethod}},
  \code{\link{isCharacterOrNULL}},
  \code{\link{isClassOrNULL}}, \code{\link{isClass}},
  \code{\link{isCountOrNULL}},
  \code{\link{isCountsOrZeros}},
  \code{\link{isFlagOrNULL}}, \code{\link{isNumberOrNULL}},
  \code{\link{isStringOrNULL}},
  \code{\link{isTRUEorFALSE}},
  \code{\link{objectHasPlaceholder}},
  \code{\link{stringMatches}},
  \code{\link{validateFrequencyCutoffs}}
}
\concept{assert functions}
