% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_utils.R
\name{isFlagOrNULL}
\alias{isFlagOrNULL}
\title{Check if object is flag or NULL}
\usage{
isFlagOrNULL(x)
}
\arguments{
\item{x}{object to test.}
}
\value{
Logical indicating if object is flag or NULL
}
\description{
\code{isFlagOrNULL} checks if object is flag (a length one logical vector) or
NULL.
}
\seealso{
Other assert functions: \code{\link{characterMatches}},
  \code{\link{checkAdditionalData}},
  \code{\link{checkHlaCallsFormat}},
  \code{\link{checkKirCountsFormat}},
  \code{\link{checkStatisticalModel}},
  \code{\link{colnamesMatches}},
  \code{\link{hasTidyMethod}},
  \code{\link{isCharacterOrNULL}},
  \code{\link{isClassOrNULL}}, \code{\link{isCountOrNULL}},
  \code{\link{isCountsOrZeros}},
  \code{\link{isNumberOrNULL}},
  \code{\link{isStringOrNULL}},
  \code{\link{isTRUEorFALSE}}, \code{\link{stringMatches}}
}
\concept{assert functions}
