% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_utils.R
\name{isInheritanceModelApplicable}
\alias{isInheritanceModelApplicable}
\title{Check if experiment is inheritance model applicable}
\usage{
isInheritanceModelApplicable(x)
}
\arguments{
\item{x}{String giving experiment type.}
}
\value{
Logical indicating if inheritance model can be applied.
}
\description{
Helper function checking if for given experiment type inheritance model can
be applied (eg. "hla_alleles") or not (eg. "hla_kir_interactions")
}
\seealso{
Other assert functions: \code{\link{characterMatches}},
  \code{\link{checkAdditionalData}},
  \code{\link{checkColDataFormat}},
  \code{\link{checkHlaCallsFormat}},
  \code{\link{checkKirCallsFormat}},
  \code{\link{checkStatisticalModel}},
  \code{\link{colnamesMatches}},
  \code{\link{functionExists}},
  \code{\link{hasTidyMethod}},
  \code{\link{isCharacterOrNULL}},
  \code{\link{isClassOrNULL}}, \code{\link{isClass}},
  \code{\link{isCountOrNULL}},
  \code{\link{isCountsOrZeros}},
  \code{\link{isExperimentCountsOrZeros}},
  \code{\link{isFlagOrNULL}}, \code{\link{isNumberOrNULL}},
  \code{\link{isStringOrNULL}},
  \code{\link{isTRUEorFALSE}},
  \code{\link{objectHasPlaceholder}},
  \code{\link{stringMatches}},
  \code{\link{validateFrequencyCutoffs}}
}
\concept{assert functions}
