% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_asserts.R
\name{isTRUEorFALSE}
\alias{isTRUEorFALSE}
\title{Check if object is TRUE or FALSE flag}
\usage{
isTRUEorFALSE(x)
}
\arguments{
\item{x}{object to test.}
}
\value{
Logical indicating if object is TRUE or FALSE flag
}
\description{
\code{isTRUEorFALSE} check if object is a flag (a length one logical vector)
except NA.
}
\seealso{
Other assert functions: \code{\link{checkColDataFormat}},
  \code{\link{hasTidyMethod}}, \code{\link{isClassOrNULL}},
  \code{\link{isCountOrNULL}},
  \code{\link{isExperimentCountsOrZeros}},
  \code{\link{objectHasPlaceholder}}
}
\concept{assert functions}
