% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_utils.R
\name{iterativeModel}
\alias{iterativeModel}
\title{Iteratively evaluate model for different variables}
\usage{
iterativeModel(object, placeholder, variables, exponentiate = FALSE)
}
\arguments{
\item{object}{An existing fit from a model function such as lm, glm and many
others.}

\item{placeholder}{String specifying term to substitute with value from
\code{x}. If more than one variable is passed in \code{x} all of them will
be bracketed. Ignored if set to \code{NULL}.}

\item{variables}{Character vector specifying variables to use in association
tests.}

\item{exponentiate}{Logical flag indicating whether or not to exponentiate
the coefficient estimates. Internally this is passed to
\code{\link[broom]{tidy}}. This is typical for logistic and multinomial
regressions, but a bad idea if there is no log or logit link. Defaults to
FALSE.}
}
\value{
Tibble containing per variable summarised statistics.
}
\description{
Iteratively evaluate model for different variables
}
