% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_transformationFunctions.R
\name{kableResults}
\alias{kableResults}
\title{Pretty format association analysis results}
\usage{
kableResults(results, colnames = NULL,
  header = "MiDAS analysis results", pvalue_cutoff = NULL,
  format = getOption("knitr.table.format"))
}
\arguments{
\item{results}{Tibble as returned by \code{\link{analyzeAssociations}}.}

\item{colnames}{Character vector giving new colnames in dply::rename style.}

\item{header}{String specifying results table header.}

\item{pvalue_cutoff}{Number specifying p-value cutoff for results to be
included in output. If \code{NULL} no filtering is done.}

\item{format}{String with possible values \code{"latex"} and \code{"html"}.}
}
\value{
A character vector with pretty formatted \code{results} table.
}
\description{
\code{kableResults} formats results table to specified format. It uses
\code{\link{formatResults}} with pre specified arguments to return pretty
formatted table depending on the type of analysis and model type.
}
\seealso{
\code{\link{formatResults}}, \code{\link{runMiDAS}}
}
