% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kir_frequencies}
\alias{kir_frequencies}
\title{KIR genes frequencies scraped from allelefrequencies.net}
\format{A data frame with 3744 rows and 3 variables:
\describe{
  \item{var}{allele number, character}
  \item{population}{reference population name, character}
  \item{frequency}{KIR genes carrier frequency in reference population, float}
}}
\source{
\url{www.allelefrequencies.net}
}
\usage{
kir_frequencies
}
\description{
Accessed on 28.08.20
}
\details{
A dataset containing KIR genes frequencies across 16 genes.
For details visit the search results page in the allelefrequencies.net
database website.
}
\keyword{datasets}
