% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_utils.R
\name{listMiDASDictionaries}
\alias{listMiDASDictionaries}
\title{List HLA alleles dictionaries}
\usage{
listMiDASDictionaries(full.names = FALSE)
}
\arguments{
\item{full.names}{Logical value. If FALSE, only the names of dictionaries are
returned. If TRUE thier file names are returned.}
}
\value{
Character vector with names of available HLA alleles dictionaries.
}
\description{
\code{listMiDASDictionaries} lists dictionaries shipped with MiDAS package.
}
