% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_raster_filename.R
\name{make_raster_filename}
\alias{make_raster_filename}
\title{make raster files}
\usage{
make_raster_filename(dekada_date = Sys.Date())
}
\arguments{
\item{dekada_date}{a character object (a date in format "YYYY/MM/DD")}
}
\value{
a list object of four names directories
}
\description{
make raster files to save it in a specific directory
}
\examples{
make_raster_filename(dekada_date = "2017-01-11")
make_raster_filename(dekada_date = "2017-01-21")
make_raster_filename(dekada_date = "2017-02-01")
}
\author{
Adrian Huerta
}
