% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{matrix_summary_plots}
\alias{matrix_summary_plots}
\title{Matrix Summary Plots}
\usage{
matrix_summary_plots(which_pool, pool_data)
}
\arguments{
\item{which_pool}{Pool number of analyze}

\item{pool_data}{tibble containing pool-wise results from \code{\link{correct_pool}}}
}
\value{
grob
}
\description{
Create heatmaps of raw abundances, the PCA projection, and corrected abundances for a given pool
}
