% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midas.R
\docType{data}
\name{metabolites}
\alias{metabolites}
\title{MIDAS Metabolites}
\format{
A tibble with 400 rows and 8 columns:
\describe{
  \item{metabolite}{Metabolite name (indistinguishable metabolites are lumped) matching \code{midas_pmis}}
  \item{metabolite_component}{A unique metabolite constituent of the \code{metabolite}}
  \item{MIDAS_ID}{\code{metabolite_component}'s internal identifier}
  \item{Pool}{Metabolite pool \code{metabolite}/\code{metabolite_component} were run in}
  \item{KEGG_ID}{\code{metabolite_component}'s KEGG identifier}
  \item{HMDB_ID}{\code{metabolite_component}'s HMDB identifier}
  \item{SMILES}{\code{metabolite_component}'s isomeric SMILES encoding structure}
  \item{KEGG_pathway_association}{\code{metabolite_component}'s KEGG pathways}
}
}
\usage{
metabolites
}
\description{
Metabolite metadata matching MIDAS measurements. Some measurements include
multiple ambiguous compounds.
}
\keyword{datasets}
