% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_class.R
\docType{class}
\name{MiDAS-class}
\alias{MiDAS-class}
\alias{MiDAS}
\alias{getInheritanceModel}
\alias{getInheritanceModel,MiDAS-method}
\alias{getExperiments}
\alias{getExperiments,MiDAS-method}
\alias{getHlaCalls}
\alias{getHlaCalls,MiDAS-method}
\alias{getKirCalls}
\alias{getKirCalls,MiDAS-method}
\alias{getPlaceholder}
\alias{getPlaceholder,MiDAS-method}
\alias{getFrequencies}
\alias{getFrequencies,MiDAS-method}
\alias{filterByFrequency}
\alias{filterByFrequency,MiDAS-method}
\title{MiDAS class}
\usage{
getInheritanceModel(object)

\S4method{getInheritanceModel}{MiDAS}(object)

getExperiments(object)

\S4method{getExperiments}{MiDAS}(object)

getHlaCalls(object)

\S4method{getHlaCalls}{MiDAS}(object)

getKirCalls(object)

\S4method{getKirCalls}{MiDAS}(object)

getPlaceholder(object)

\S4method{getPlaceholder}{MiDAS}(object)

getFrequencies(object, experiment)

\S4method{getFrequencies}{MiDAS}(object, experiment)

filterByFrequency(object, experiment, lower_frequency_cutoff,
  upper_frequency_cutoff)

\S4method{filterByFrequency}{MiDAS}(object, experiment,
  lower_frequency_cutoff = NULL, upper_frequency_cutoff = NULL)
}
\arguments{
\item{object}{\code{\link{MiDAS}} object}

\item{experiment}{String specifying experiment.}

\item{lower_frequency_cutoff}{Number}

\item{upper_frequency_cutoff}{Number}

\item{object}{\code{\link{MiDAS}} object}

\item{object}{\code{\link{MiDAS}} object}

\item{object}{\code{\link{MiDAS}} object.}

\item{object}{\code{\link{MiDAS}} object.}

\item{object}{\code{\link{MiDAS}} object.}

\item{object}{\code{\link{MiDAS}} object}

\item{experiment}{Matrix}
}
\value{
String name of object's inheritance model.

Character giving names of available analysis types.

Data frame with object's HLA calls.

Data frame with object's KIR calls.

String giving object's placeholder.

Data frame containing variables, its corresponding total counts
  and frequencies.

Variables frequencies are counted in reference to sample size, depending on
the inheritance model under which the counts table has been generated one
might need to take under consideration both gene copies. Here sample size is
assumed to be depended on both gene copies for \code{"additive"} inheritance
model (`n / (2 * j)` where `n` is the number of term occurrences and `j`
is the sample size). For other models the sample size is taken as is
(`n / j`).

Object of class MiDAS.
}
\description{
The \code{MiDAS} class is a \code{\link{MultiAssayExperiment}} object
containing all the data and metadata about a set of measurments and thier
transformations required for MiDAS analysis.
}
\details{
A object of class \code{MiDAS} have at least one of
\link[=readHlaCalls]{HLA calls} or \link[=readKirCalls]{KIR calls} experiment
matrices defined (they can be accessed using \code{\link{getHlaCalls}} and
\code{\link{getKirCalls}} functions).

Other experiments slots can be occupied by transformations of those matrices.
Those transformations can be further used by \code{\link{runMiDAS}} function
for performing statistical analyses. See \code{\link{prepareMiDAS}} and
\code{\link{runMiDAS}} for more informations.

An inheritance model of \code{MiDAS} object is held in object's metadata. It
specifies inheritance model under which all experiments matrices has been
constructed. It can be accessed using \code{\link{getInheritanceModel}}
function.

Experiments slots available for \code{\link{runMiDAS}} function are defined
in object's metadata under \code{experiment} variable. It can be accessed
using \code{\link{getExperiments}} function.
}
