% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_class.R
\name{MiDAS}
\alias{MiDAS}
\title{MiDAS class constructor}
\usage{
MiDAS(hla_calls = NULL, kir_counts = NULL, colData = NULL,
  inheritance_model = c("additive", "dominant", "recesive"),
  analysis_type = c("hla_allele", "aa_level", "expression_level",
  "allele_g_group", "allele_supertype", "allele_group", "kir_genes",
  "hla_kir_interactions"))
}
\description{
MiDAS class constructor
}
