% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{midas_barplot}
\alias{midas_barplot}
\title{MIDAS Barplot}
\usage{
midas_barplot(midas_data, facet_by = "protein", FDR_cutoff = 0.1)
}
\arguments{
\item{midas_data}{\code{\link{midas_pmis}} with any filters applied}

\item{facet_by}{Variable to facet by, either protein or metabolite}

\item{FDR_cutoff}{cutoff used for labeling discoveries}
}
\value{
a grob
}
\description{
MIDAS Barplot
}
\examples{

library(dplyr)
library(ggplot2)

midas_pmis \%>\%
  dplyr::filter(query_protein \%in\%
    sample(unique(midas_pmis$query_protein), 3)) \%>\%
  midas_barplot(FDR_cutoff = 0.1) +
  theme(axis.text.x = element_blank())

}
