% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midas.R
\docType{data}
\name{midas_pmis}
\alias{midas_pmis}
\title{MIDAS PMIs}
\format{
A tibble with 52,614 rows and 8 columns:
\describe{
  \item{metabolite}{Metabolite name (indistinguishable metabolites are lumped)}
  \item{query_protein}{Protein name. Anonymous proteins are named MIDASXXXX.}
  \item{log2_abundance}{Raw fold-changes between protein-present and protein-absent compartments}
  \item{log2_abundance_corrected}{Fold-changes corrected for non-specific MPIs using PCA}
  \item{p_value}{P-value for enrichment/depletion}
  \item{q_value}{P-values corrected for multiple hypotheses using PMID: 12883005 }
}
}
\usage{
midas_pmis
}
\description{
MIDAS Protein-Metabolite Interactions (PMIs) produced by running the vignette, pipeline.Rmd.
}
\keyword{datasets}
