% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midas_touch.R
\name{midas_touch}
\alias{midas_touch}
\title{turns an xml2 list into a shiny function call}
\usage{
midas_touch(root)
}
\arguments{
\item{root}{an html list (from xml2 as_list).
Note: If the list is particularly deep,
you may need to set option(expressions = SOMETHING BIG)}
}
\value{
a function call that produces the equivalent shiny objects
}
\description{
turns an xml2 list into a shiny function call
}
\examples{
html <- '<div class="example"><h3>test</h3></div>'
html_list <- html_to_list(html)
gold <- midas_touch(html_list)
eval(gold)
}

