% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MiDAS_tut_HLA}
\alias{MiDAS_tut_HLA}
\title{MiDAS tutorial HLA data}
\format{Data frame with 1000 rows and 19 columns. First column holds samples
  ID's, following columns holds HLA alleles calls for different genes.}
\usage{
MiDAS_tut_HLA
}
\description{
Example HLA calls data used in MiDAS tutorial
}
\keyword{datasets}
