% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MiDAS_tut_KIR}
\alias{MiDAS_tut_KIR}
\title{MiDAS tutorial KIR data}
\format{Data frame with 1000 rows and 17 columns. First column holds samples
  ID's, following columns holds presence/absence indicators for different
  KIR genes.}
\usage{
MiDAS_tut_KIR
}
\description{
Example KIRR presence/absence data used in MiDAS tutorial
}
\keyword{datasets}
