% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_utils.R
\name{objectHasPlaceholder}
\alias{objectHasPlaceholder}
\title{Check if placeholder is present in object formula}
\usage{
objectHasPlaceholder(object, placeholder)
}
\arguments{
\item{object}{statistical model to test.}

\item{placeholder}{string specifying name of placeholder.}
}
\value{
Logical indicating if placeholder is present in object formula.
}
\description{
\code{isTRUEorFALSE} check if object is a flag (a length one logical vector)
except NA.
}
\seealso{
Other assert functions: \code{\link{characterMatches}},
  \code{\link{checkAdditionalData}},
  \code{\link{checkColDataFormat}},
  \code{\link{checkHlaCallsFormat}},
  \code{\link{checkKirCallsFormat}},
  \code{\link{checkStatisticalModel}},
  \code{\link{colnamesMatches}},
  \code{\link{functionExists}},
  \code{\link{hasTidyMethod}},
  \code{\link{isCharacterOrNULL}},
  \code{\link{isClassOrNULL}}, \code{\link{isClass}},
  \code{\link{isCountOrNULL}},
  \code{\link{isCountsOrZeros}},
  \code{\link{isExperimentCountsOrZeros}},
  \code{\link{isFlagOrNULL}},
  \code{\link{isInheritanceModelApplicable}},
  \code{\link{isNumberOrNULL}},
  \code{\link{isStringOrNULL}},
  \code{\link{isTRUEorFALSE}}, \code{\link{stringMatches}},
  \code{\link{validateFrequencyCutoffs}}
}
\concept{assert functions}
