% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dekadalstat_amount.R
\name{plot_dekadalstat_amount}
\alias{plot_dekadalstat_amount}
\title{plot of dekadal stat amount}
\usage{
plot_dekadalstat_amount(r = drougth_classification(),
  dir_save_file = "C:/Users/usuario 01/Desktop/DEK_files",
  name_file = make_raster_filename()$raster_dek_png)
}
\arguments{
\item{r}{a raster object}

\item{dir_save_file}{a character object (directory where is saved)}

\item{name_file}{a character object}
}
\value{
a png file
}
\description{
plot of dekadal stat amount and save as a png file
}
\author{
Adrian Huerta & Carlos Fernandez
}
