% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_stats.R
\name{prepareHlaData}
\alias{prepareHlaData}
\title{Prepare data for statistical analysis}
\usage{
prepareHlaData(hla_calls, pheno, covar = NULL,
  inheritance_model = c("dominant", "recessive", "additive"))
}
\arguments{
\item{hla_calls}{Data frame containing HLA allele calls, as return by
\code{\link{readHlaCalls}} function.}

\item{pheno}{Data frame holding phenotypic response variables.}

\item{covar}{Data frame holding covariates.}

\item{inheritance_model}{String specifying inheritance model to use.
Available choices are \code{"dominant"}, \code{"recessive"},
\code{"additive"}. In \code{"dominant"} model dominant homozygotes and
dominant heterozygotes are coded as \code{1} and recesive homozygotes as
\code{0}. In \code{"recessive"} model dominant homozygotes are coded as
\code{1} and all other as \code{0}. In \code{"additive"} model dominant
homozygotes are coded as \code{2}, dominant heterozygotes as \code{1} and
recesive homozygotes as \code{0}.}
}
\description{
Prepare data for statistical analysis
}
\examples{
hla_calls_file <- system.file("extdata", "HLAHD_output_example.txt", package = "MiDAS")
hla_calls <- readHlaCalls(hla_calls_file)
pheno_file <- system.file("extdata", "pheno_example.txt", package = "MiDAS")
pheno <- read.table(pheno_file, header = TRUE)
covar_file <- system.file("extdata", "covar_example.txt", package = "MiDAS")
covar <- read.table(covar_file, header = TRUE)
prepareHlaData(hla_calls = hla_calls,
               pheno = pheno,
               covar = covar,
               inheritance_model = "additive"
)

}
