% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_class.R
\name{prepareMiDAS}
\alias{prepareMiDAS}
\alias{prepareMiDAS_hla_alleles}
\title{Prepare MiDAS object for statistical analysis}
\usage{
prepareMiDAS(hla_calls = NULL, kir_calls = NULL, colData,
  experiment = c("hla_alleles", "hla_aa", "hla_g_groups",
  "hla_supertypes", "hla_NK_ligands", "kir_genes", "hla_kir_interactions",
  "hla_divergence"), placeholder = "term",
  lower_frequency_cutoff = NULL, upper_frequency_cutoff = NULL, ...)

prepareMiDAS_hla_alleles(hla_calls, ...)
}
\arguments{
\item{hla_calls}{Data frame containing HLA allele calls, as return by
\code{\link{readHlaCalls}} function.}

\item{kir_calls}{Data frame containing KIR gene counts, as returned by
\code{\link{readKPICalls}} function.}

\item{colData}{Data frame holding additional variables like phenotypic
observations or covariates. It have to contain \code{'ID'} column holding
samples identifiers corresponding to identifiers in \code{hla_calls} and
\code{kir_calls}. Importantly rows of \code{hla_calls} and
\code{kir_calls} without corresponding phenotype are discarded.}

\item{experiment}{Character vector indicating analysis type for which data
should be prepared. Valid choices are \code{"hla_alleles"},
\code{"hla_aa"}, \code{"hla_g_groups"}, \code{"hla_supertypes"},
\code{"hla_NK_ligands"}, \code{"kir_genes"}, \code{"hla_kir_interactions"},
\code{"hla_divergence"}.
See details for further explanations.}

\item{placeholder}{String}

\item{lower_frequency_cutoff}{Number}

\item{upper_frequency_cutoff}{Number}

\item{...}{Attributes used in data transformation.}
}
\value{
Object of class \code{\link{MiDAS}}

Matrix
}
\description{
\code{prepareMiDAS} transform HLA alleles calls and KIR calls according
to selected analysis creating \code{\link{MiDAS}} object.
}
\details{
\code{'hla_alleles'} - \code{hla_calls} are transformed into counts.
}
\examples{
\dontrun{
# read hla calls file
hla_calls_file <- system.file("extdata", "HLAHD_output_example.txt", package = "MiDAS")
hla_calls <- readHlaCalls(hla_calls_file)

# read kir calls file
kir_calls_file <- system.file("extdata", "KPI_output_example.txt", package = "MiDAS")
kir_calls <- readKPICalls(kir_calls_file, counts = TRUE)

# read phenotypic data and covariates
pheno_file <- system.file("extdata", "pheno_example.txt", package = "MiDAS")
pheno <- read.table(pheno_file, header = TRUE, stringsAsFactors = FALSE)
covar_file <- system.file("extdata", "covar_example.txt", package = "MiDAS")
covar <- read.table(covar_file, header = TRUE, stringsAsFactors = FALSE)
phenotype <- left_join(pheno, covar, by ="ID")

# create MiDAS object
midas <- prepareMiDAS(hla_calls = hla_calls,
                      kir_calls = kir_calls,
                      colData = phenotype,
                      experiment = "hla_alleles"
)
}

}
