% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_class.R
\name{prepareMiDAS_aa_level}
\alias{prepareMiDAS_aa_level}
\title{Prepare MiDAS data on HLA amino acid level}
\usage{
prepareMiDAS_aa_level(hla_calls, inheritance_model, indels = TRUE,
  unkchar = FALSE, ...)
}
\arguments{
\item{hla_calls}{Data frame}

\item{inheritance_model}{String}

\item{...}{Not used}
}
\value{
Matrix
}
\description{
Prepare MiDAS data on HLA amino acid level
}
