% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_class.R
\name{prepareMiDAS_allele_group}
\alias{prepareMiDAS_allele_group}
\title{Prepare MiDAS data on HLA allele's groups level}
\usage{
prepareMiDAS_allele_group(hla_calls, inheritance_model, ...)
}
\arguments{
\item{hla_calls}{Data frame}

\item{inheritance_model}{String}

\item{...}{Not used}
}
\value{
Matrix
}
\description{
\code{hla_calls} are transformed to HLA alleles groups using Bw4/6, C1/2 and
Bw4+A23+A24+A32 dictionaries shipped with the package. Than they are
transformed to counts under \code{inheritance_model} of choice (see
\code{\link{hlaToVariable}} and \code{\link{hlaCallsToCounts}} for more
details).
}
