% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_class.R
\name{prepareMiDAS_allele_supertype}
\alias{prepareMiDAS_allele_supertype}
\title{Prepare MiDAS data on HLA allele's supertypes level}
\usage{
prepareMiDAS_allele_supertype(hla_calls, inheritance_model, ...)
}
\arguments{
\item{hla_calls}{Data frame}

\item{inheritance_model}{String}

\item{...}{Not used}
}
\value{
Matrix
}
\description{
Prepare MiDAS data on HLA allele's supertypes level
}
