% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_class.R
\name{prepareMiDAS_hla_allele}
\alias{prepareMiDAS_hla_allele}
\title{Prepare MiDAS data on HLA allele level}
\usage{
prepareMiDAS_hla_allele(hla_calls, inheritance_model, ...)
}
\arguments{
\item{hla_calls}{Data frame}

\item{inheritance_model}{String}

\item{...}{Not used}
}
\value{
Matrix
}
\description{
Prepare MiDAS data on HLA allele level
}
