% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_class.R
\name{prepareMiDAS_hla_divergence}
\alias{prepareMiDAS_hla_divergence}
\title{Prepare MiDAS data on HLA divergence level}
\usage{
prepareMiDAS_hla_divergence(hla_calls, ...)
}
\arguments{
\item{hla_calls}{Data frame}

\item{...}{Not used}
}
\value{
Matrix
}
\description{
Distances between Class I alleles are calculated using Grantham distance,
as implemented in \code{hlaCallsGranthamDistance} function. Additionally
average distance in Class I genese is calculated.
}
