% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_class.R
\name{prepareMiDAS_hla_g_groups}
\alias{prepareMiDAS_hla_g_groups}
\title{Prepare MiDAS data on HLA allele's G groups level}
\usage{
prepareMiDAS_hla_g_groups(hla_calls, ...)
}
\arguments{
\item{hla_calls}{HLA calls data frame, as returned by
\code{\link{readHlaCalls}} function.}

\item{...}{Not used}
}
\value{
Matrix
}
\description{
Prepare MiDAS data on HLA allele's G groups level
}
