% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_class.R
\name{prepareMiDAS_hla_het}
\alias{prepareMiDAS_hla_het}
\title{Prepare MiDAS data on HLA heterozygosity level}
\usage{
prepareMiDAS_hla_het(hla_calls, hla_het_resolution = 8, ...)
}
\arguments{
\item{hla_calls}{Data frame}

\item{...}{Not used}
}
\value{
Matrix
}
\description{
Heterozygosity status for each allele, "1" designates homozygote and "0"
designates heterozygote.
}
