% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_class.R
\name{prepareMiDAS_hla_NK_ligands}
\alias{prepareMiDAS_hla_NK_ligands}
\title{Prepare MiDAS data on HLA allele's groups level}
\usage{
prepareMiDAS_hla_NK_ligands(hla_calls, ...)
}
\arguments{
\item{hla_calls}{Data frame}

\item{...}{Not used}
}
\value{
Matrix
}
\description{
\code{hla_calls} are transformed to HLA alleles groups using Bw4/6, C1/2 and
Bw4+A23+A24+A32 dictionaries shipped with the package. Than they are
transformed to counts.
}
