% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_class.R
\name{prepareMiDAS_hla_supertypes}
\alias{prepareMiDAS_hla_supertypes}
\title{Prepare MiDAS data on HLA allele's supertypes level}
\usage{
prepareMiDAS_hla_supertypes(hla_calls, ...)
}
\arguments{
\item{hla_calls}{Data frame}

\item{...}{Not used}
}
\value{
Matrix
}
\description{
\code{hla_calls} are transformed to HLA alleles groups using supertypes
dictionary shipped with the package. Than they are transformed to counts).
}
