% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_class.R
\name{prepareMiDAS_kir_haplotypes}
\alias{prepareMiDAS_kir_haplotypes}
\title{Prepare MiDAS data on KIR haplotypes level}
\usage{
prepareMiDAS_kir_haplotypes(kir_calls, ...)
}
\arguments{
\item{kir_calls}{KIR calls data frame, as returned by
\code{\link{readKIRCalls}} function.}
}
\value{
Matrix
}
\description{
Prepare MiDAS data on KIR haplotypes level
}
