% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_utils.R
\name{procHlaAlignmentGrantham}
\alias{procHlaAlignmentGrantham}
\title{Helper function processing alignment for Grantham distance calculations}
\usage{
procHlaAlignmentGrantham(alignment)
}
\arguments{
\item{alignment}{Matrix containing HLA alleles amino acids level alignment.}
}
\description{
\code{procHlaAlignmentGrantham} apply processing to raw HLA alignments, so
that grantham distance between alleles can be calculated. Processing includes
extracting exons 1 and 2, masking indels, gaps and stop codons.
}
