% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_parsingFunctions.R
\name{readKIRCalls}
\alias{readKIRCalls}
\title{Read KIR calls}
\usage{
readKIRCalls(file, na.strings = c("", "NA", "uninterpretable"))
}
\arguments{
\item{file}{Path to input file.}

\item{na.strings}{a character vector of strings which are to be
    interpreted as \code{\link{NA}} values.  Blank fields are also
    considered to be missing values in logical, integer, numeric and
    complex fields.  Note that the test happens \emph{after} 
    white space is stripped from the input, so \code{na.strings} 
    values may need their own white space stripped in advance.}
}
\value{
Data frame containing KIR gene's counts.
}
\description{
\code{readKiRCalls} read KIR calls from file.
}
\details{
Input file has to be a tsv formatted table. First column should be named
"ID" and contain samples IDs, second column should hold corresponding KIR
haplotypes, further columns should hold KIR genes presence / absence
indicators. See
\code{system.file("extdata", "KPI_output_example.txt", package = "MiDAS")}
for an example.
}
\examples{
file <- system.file("extdata", "KPI_output_example.txt", package = "MiDAS")
readKIRCalls(file)

}
