% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_transformationFunctions.R
\name{reduceHlaCalls}
\alias{reduceHlaCalls}
\title{Reduce HLA calls data frame resolution}
\usage{
reduceHlaCalls(hla_calls, resolution = 4)
}
\arguments{
\item{hla_calls}{Data frame containing HLA allele calls, as return by
\code{\link{readHlaCalls}} function.}

\item{resolution}{Numeric vector of length one specifying output resolution.}
}
\value{
Data frame containing HLA allele calls reduced to required
  resolution.
}
\description{
\code{reduceHlaCalls} reduce HLA calls data frame to specified resolution.
}
\details{
If \code{resolution} is greater than resolution of \code{hla_calls} elements,
those elements will be unchanged. Elements with optional suffixes are not
reduced.
}
\examples{
file <- system.file("extdata", "HLAHD_output_example.txt", package = "MiDAS")
hla_calls <- readHlaCalls(file)
reduceHlaCalls(hla_calls, resolution = 2)

}
