% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_utils.R
\name{runMiDASGetVarsFreq}
\alias{runMiDASGetVarsFreq}
\title{Add variables frequencies to runMiDAS results}
\usage{
runMiDASGetVarsFreq(midas, analysis_type, test_covar)
}
\arguments{
\item{midas}{MiDAS object}

\item{test_covar}{String giving name of test covariate}
}
\description{
This is a helper function
}
