% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_stats.R
\name{runMiDASMode<-}
\alias{runMiDASMode<-}
\alias{runMiDASMode}
\title{Helper for defining runMiDAS modes}
\usage{
runMiDASMode(x, mode) <- value

runMiDASMode(x, mode)
}
\arguments{
\item{x}{runMiDAS mode function}

\item{which}{String attribute name}
}
\description{
Helper for defining runMiDAS modes

Helper for accessing runMiDAS modes
}
