% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_utils.R
\name{subsetMiDASByFreq}
\alias{subsetMiDASByFreq}
\title{Subset MiDAS data by frequency}
\usage{
subsetMiDASByFreq(midas_data, lower_frequency_cutoff = NULL,
  upper_frequency_cutoff = NULL)
}
\arguments{
\item{upper_frequency_cutoff}{}
}
\description{
\code{subsetMiDASByFreq} subsets \code{midas_data} selecting variables that
meet frequency criteria. Only variables of "integer" type are considered for
filtration, variables of other types (eg. "float") are returend unaffected.
}
\details{
TODO write unit tests
}
\examples{
hla_calls_file <- system.file("extdata", "HLAHD_output_example.txt", package = "MiDAS")
hla_calls <- readHlaCalls(hla_calls_file)
pheno_file <- system.file("extdata", "pheno_example.txt", package = "MiDAS")
pheno <- read.table(pheno_file, header = TRUE, stringsAsFactors = FALSE)
covar_file <- system.file("extdata", "covar_example.txt", package = "MiDAS")
covar <- read.table(covar_file, header = TRUE, stringsAsFactors = FALSE)
midas_data <- prepareMiDAS(hla_calls = hla_calls,
                           pheno = pheno,
                           covar = covar,
                           analysis_type = "hla_allele",
                           inheritance_model = "additive"
)
subsetMiDASByFreq(midas_data,
                  lower_frequency_cutoff = 0.1,
                  upper_frequency_cutoff = 0.9
)


}
