% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_utils.R
\name{validateFrequencyCutoffs}
\alias{validateFrequencyCutoffs}
\title{Validate frequency cutoffs}
\usage{
validateFrequencyCutoffs(lower_frequency_cutoff, upper_frequency_cutoff)
}
\arguments{
\item{lower_frequency_cutoff}{Number}

\item{upper_frequency_cutoff}{Number}
}
\value{
Logical indicating if \code{lower_frequency_cutoff} and
  \code{upper_frequency_cutoff} are valid.
}
\description{
\code{validateFrequencyCutoffs} checks if \code{lower_frequency_cutoff} and
\code{upper_frequency_cutoff} are valid.
}
\details{
\code{lower_frequency_cutoff} and \code{upper_frequency_cutoff} should be a
positive numbers, giving either frequency or counts.
\code{lower_frequency_cutoff} has to be lower than
\code{upper_frequency_cutoff}.
}
\seealso{
Other assert functions: \code{\link{characterMatches}},
  \code{\link{checkAdditionalData}},
  \code{\link{checkColDataFormat}},
  \code{\link{checkHlaCallsFormat}},
  \code{\link{checkKirCallsFormat}},
  \code{\link{checkStatisticalModel}},
  \code{\link{colnamesMatches}},
  \code{\link{functionExists}},
  \code{\link{hasTidyMethod}},
  \code{\link{isCharacterOrNULL}},
  \code{\link{isClassOrNULL}}, \code{\link{isClass}},
  \code{\link{isCountOrNULL}},
  \code{\link{isCountsOrZeros}},
  \code{\link{isExperimentCountsOrZeros}},
  \code{\link{isFlagOrNULL}},
  \code{\link{isInheritanceModelApplicable}},
  \code{\link{isNumberOrNULL}},
  \code{\link{isStringOrNULL}},
  \code{\link{isTRUEorFALSE}},
  \code{\link{objectHasPlaceholder}},
  \code{\link{stringMatches}}
}
\concept{assert functions}
