% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_transformationFunctions.R
\name{aaVariationToCounts}
\alias{aaVariationToCounts}
\title{Transform amino acids variation data frame to counts table}
\usage{
aaVariationToCounts(aa_variation, inheritance_model = c("dominant",
  "recessive", "additive"))
}
\arguments{
\item{aa_variation}{Data frame holding amino acid variation data as returned
by \link{hlaToVariable}.}

\item{inheritance_model}{String specifying inheritance model to use.
Available choices are \code{"dominant"}, \code{"recessive"},
\code{"additive"}. In \code{"dominant"} model homozygotes and heterozygotes
are coded as \code{1}. In \code{"recessive"} model homozygotes are coded as
\code{1} and all other as \code{0}. In \code{"additive"} model homozygotes
are coded as \code{2} and heterozygotes as \code{1}.}
}
\value{
Data frame containing counts of amino acids at specific positions
  counted according to specified model.
}
\description{
\code{aaVariationToCounts} converts variation data frame data frame into
counts table.
}
\examples{
file <- system.file("extdata", "HLAHD_output_example.txt", package = "MiDAS")
hla_calls <- readHlaCalls(file)
aa_variation <- hlaToAAVariation(hla_calls)
aaVariationToCounts(aa_variation, inheritance_model = "additive")

}
