% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_transformationFunctions.R
\name{aaVariationToCounts}
\alias{aaVariationToCounts}
\title{Transform amino acid variation data frame into counts table}
\usage{
aaVariationToCounts(aa_variation)
}
\arguments{
\item{aa_variation}{Amino acid variation data frame as returned by
\link{hlaToAAVariation}.}
}
\value{
Amino acid counts data frame.
}
\description{
\code{aaVariationToCounts} convert amino acid variation data frame into
counts table.
}
