% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{allele_frequencies}
\alias{allele_frequencies}
\title{Alleles frequencies scraped from allelefrequencies.net}
\format{
A data frame with 2096 rows and 3 variables:
\describe{
  \item{var}{allele number, character}
  \item{population}{reference population name, character}
  \item{frequency}{allele frequency in reference population, float}
}
}
\source{
\url{www.allelefrequencies.net}
}
\usage{
allele_frequencies
}
\description{
Accessed on 28.07.20
}
\details{
A dataset containing allele frequencies across 5697 alleles
For details visit the search results page in the allelefrequencies.net
database website.
}
\keyword{datasets}
