% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_stats.R
\name{analyzeAssociations}
\alias{analyzeAssociations}
\title{Association analysis}
\usage{
analyzeAssociations(object, variables, correction = "bonferroni",
  exponentiate = FALSE)
}
\arguments{
\item{object}{An existing fit from a model function such as lm, glm and many
others.}

\item{variables}{Character specifying variables to use in association tests.}

\item{correction}{String specifying multiple testing correction method. See
details for further information.}

\item{exponentiate}{Logical indicating whether or not to exponentiate the
  coefficient estimates. Internally this is passed to \link[broom]{tidy}.
  This is typical for logistic and multinomial regressions, but a bad idea if
  there is no log or logit link. Defaults to FALSE.

\code{correction} specifies p-value adjustment method to use, common choice
is Benjamini & Hochberg (1995) (\code{"BH"}). Internally this is passed to
\link[stats]{p.adjust}. Check there to get more details.}
}
\value{
Tibble containing combined results for all alleles in
  \code{hla_calls}.
}
\description{
\code{analyzeAssociations} performs associations analysis on single variable
level using statistical model of choice.
}
\examples{
library("survival")
hla_calls_file <- system.file("extdata", "HLAHD_output_example.txt", package = "MiDAS")
hla_calls <- readHlaCalls(hla_calls_file)
pheno_file <- system.file("extdata", "pheno_example.txt", package = "MiDAS")
pheno <- read.table(pheno_file, header = TRUE)
covar_file <- system.file("extdata", "covar_example.txt", package = "MiDAS")
covar <- read.table(covar_file, header = TRUE)
midas_data <- prepareHlaData(hla_calls = hla_calls,
                             pheno = pheno,
                             covar = covar,
                             inheritance_model = "additive"
)

# Cox proportional hazards regression model
## define base model with response and covariates
object <- coxph(Surv(OS, OS_DIED) ~ AGE + SEX, data = midas_data)

## test for alleles associations
analyzeAssociations(object = object,
                    variables = c("B*14:02", "DRB1*11:01")
)

}
