% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_stats.R
\name{analyzeMiDASData}
\alias{analyzeMiDASData}
\title{Analyze associations in MiDAS data}
\usage{
analyzeMiDASData(object, conditional = FALSE, variables = NULL,
  frequency_cutoff = NULL, pvalue_cutoff = NULL, correction = "BH",
  th = 0.05, rss_th = 1e-07, kable_output = TRUE,
  type = "hla_alleles", format = getOption("knitr.table.format"))
}
\arguments{
\item{object}{An existing fit from a model function such as lm, glm and many
others.}

\item{conditional}{Logical indicating if the analysis should be performed
using stepwise conditional tests or not. See
\link{analyzeConditionalAssociations} for more details.}

\item{variables}{Character specifying variables to use in association tests
or \code{NULL}. If \code{NULL} all variables in object data are tested.
See details for further information.}

\item{frequency_cutoff}{Number specifying threshold for inclusion of a
variable. If it's a number between 0 and 1 variables with frequency below
this number will not be considered during analysis. If it's greater or
equal 1 variables with number of counts less that this will not be
considered during analysis.}

\item{pvalue_cutoff}{Number specifying p-value cutoff for results to be
included in output. If \code{NULL} cutoff of \code{0.05} on
\code{p.adjusted} value is used instead.}

\item{correction}{String specifying multiple testing correction method. See
details for further information.}

\item{th}{number specifying p-value threshold for a variable to be considered
significant.}

\item{rss_th}{number specifying residual sum of squares threshold at which
function should stop adding additional variables. As the residual sum of
squares approaches \code{0} the perfect fit is obtained making further
attempts at variables selection nonsense, thus function is stopped. This
behavior can be controlled using \code{rss_th}.}

\item{kable_output}{Logical indicating if additionally results should be
  pretty printed in specified \code{format}.

\code{variables} takes \code{NULL} as a default value. When specified as such
column names of data frame associated with the \code{object} are used as
variables for testing. This excludes first column which should corresponds
to samples IDs as well as covariates and response variables defined in
object formula.

\code{correction} specifies p-value adjustment method to use, common choice
is Benjamini & Hochberg (1995) (\code{"BH"}). Internally this is passed to
\link[stats]{p.adjust}. Check there to get more details.}

\item{type}{String specifying type of analysis from which \code{results} were
produced. Possible values includes \code{'hla_alleles'}, \code{'aa_level'},
\code{'expression_levels'}.}

\item{format}{A character string. Possible values are \code{"latex"} and
\code{"html"}.}
}
\value{
Tibble containing results for all tested variables.
}
\description{
\code{analyzeMiDASData} performs association analysis on MiDAS data using
statistical model specified by user.
}
\examples{
library("survival")
hla_calls_file <- system.file("extdata", "HLAHD_output_example.txt", package = "MiDAS")
hla_calls <- readHlaCalls(hla_calls_file)
pheno_file <- system.file("extdata", "pheno_example.txt", package = "MiDAS")
pheno <- read.table(pheno_file, header = TRUE, stringsAsFactors = FALSE)
covar_file <- system.file("extdata", "covar_example.txt", package = "MiDAS")
covar <- read.table(covar_file, header = TRUE, stringsAsFactors = FALSE)
midas_data <- prepareHlaData(hla_calls = hla_calls,
                             pheno = pheno,
                             covar = covar,
                             inheritance_model = "additive"
)

object <- coxph(Surv(OS, OS_DIED) ~ AGE + SEX, data = midas_data)
analyzeMiDASData(object)

}
