% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_transformationFunctions.R
\name{applyInheritanceModel}
\alias{applyInheritanceModel}
\alias{applyInheritanceModel.matrix}
\alias{applyInheritanceModel.SummarizedExperiment}
\title{Format experiment to inheritance model}
\usage{
applyInheritanceModel(experiment, inheritance_model = c("dominant",
  "recessive", "additive"))

\method{applyInheritanceModel}{matrix}(experiment,
  inheritance_model = c("dominant", "recessive", "additive"))

\method{applyInheritanceModel}{SummarizedExperiment}(experiment,
  inheritance_model = c("dominant", "recessive", "additive"))
}
\arguments{
\item{experiment}{Matrix}

\item{inheritance_model}{String specifying inheritance model to use.
Available choices are \code{"dominant"}, \code{"recessive"},
\code{"additive"}. In \code{"dominant"} model homozygotes and heterozygotes
are coded as \code{1}. In \code{"recessive"} model homozygotes are coded as
\code{1} and all other as \code{0}. In \code{"additive"} model homozygotes
are coded as \code{2} and heterozygotes as \code{1}.}
}
\value{
Matrix
}
\description{
Helper function transforming experiments counts to selected
\code{inheritance_model}. Under \code{"additive"} inheritance model
experiment is returned unchanged. For \code{"dominant"} occurence of the
variable is reported as \code{1} and \code{0} otherwise. Under
\code{"recessive"} inheritence model occurence of the variable occurence of
two copies of the variable is reported as \code{1} and \code{0} otherwise.
}
