% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_utils.R
\name{backquote}
\alias{backquote}
\title{Backquote string}
\usage{
backquote(x)
}
\arguments{
\item{x}{Character vector.}
}
\value{
Character vector with its elements backticked.
}
\description{
\code{backquote} places backticks around string.
}
\details{
\code{backquote} is usefull when using HLA allele numbers in fomulas, where
\code{'*'} and \code{':'} characters have special meanings.
}
\examples{
backquote("A*01:01")

}
