% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_utils.R
\name{characterMatches}
\alias{characterMatches}
\title{Check if character matches one of possible values}
\usage{
characterMatches(x, choice)
}
\arguments{
\item{x}{Character vector to test.}

\item{choice}{Character vector with possible values for \code{x}.}
}
\value{
Logical indicating if \code{x}'s elements matches any of the values
in \code{choice}.
}
\description{
\code{characterMatches} checks if all elements of a character vector matches
values in choices.
}
\seealso{
Other assert functions: \code{\link{checkAdditionalData}},
  \code{\link{checkHlaCallsFormat}},
  \code{\link{checkKirCountsFormat}},
  \code{\link{checkStatisticalModel}},
  \code{\link{colnamesMatches}},
  \code{\link{hasTidyMethod}},
  \code{\link{isCharacterOrNULL}},
  \code{\link{isClassOrNULL}}, \code{\link{isCountOrNULL}},
  \code{\link{isCountsOrZeros}},
  \code{\link{isFlagOrNULL}}, \code{\link{isNumberOrNULL}},
  \code{\link{isStringOrNULL}},
  \code{\link{isTRUEorFALSE}},
  \code{\link{objectHasPlaceholder}},
  \code{\link{stringMatches}}
}
\concept{assert functions}
