% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_utils.R
\name{checkAlleleFormat}
\alias{checkAlleleFormat}
\title{Checks allele format}
\usage{
checkAlleleFormat(allele)
}
\arguments{
\item{allele}{Character vector containing HLA allele numbers.}

\item{resolution}{Integer vector of length one specifiying desired
resolution.}
}
\value{
Logical vector of length specifying if \code{allele} follows HLA
alleles naming conventions and have desired resolution.
}
\description{
Checks allele format
}
\examples{
allele <- c("A*01:01", "A*01:02")
checkAlleleFormat(allele)

}
