% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_utils.R
\name{checkKirCallsFormat}
\alias{checkKirCallsFormat}
\title{Assert kir calls data frame format}
\usage{
checkKirCallsFormat(kir_calls)
}
\arguments{
\item{kir_calls}{Data frame containing KIR calls, as return by
\code{\link{readKirCalls}} function.}
}
\value{
Logical indicating if \code{kir_calls} follows kir calls data frame
  format. Otherwise raise error.
}
\description{
\code{checkKirCallsFormat} asserts if kir calls data frame have proper
format.
}
\examples{
file <- system.file("extdata", "KIP_output_example.txt", package = "MiDAS")
kir_calls <- readKirCalls(file)
checkKirCallsFormat(kir_calls)

}
\seealso{
Other assert functions: \code{\link{characterMatches}},
  \code{\link{checkAdditionalData}},
  \code{\link{checkColDataFormat}},
  \code{\link{checkHlaCallsFormat}},
  \code{\link{checkKirCountsFormat}},
  \code{\link{checkStatisticalModel}},
  \code{\link{colnamesMatches}},
  \code{\link{functionExists}},
  \code{\link{hasTidyMethod}},
  \code{\link{isCharacterOrNULL}},
  \code{\link{isClassOrNULL}}, \code{\link{isClass}},
  \code{\link{isCountOrNULL}},
  \code{\link{isCountsOrZeros}},
  \code{\link{isFlagOrNULL}}, \code{\link{isNumberOrNULL}},
  \code{\link{isStringOrNULL}},
  \code{\link{isTRUEorFALSE}},
  \code{\link{objectHasPlaceholder}},
  \code{\link{stringMatches}},
  \code{\link{validateFrequencyCutoffs}}
}
\concept{assert functions}
