% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_asserts.R
\name{checkKirCallsFormat}
\alias{checkKirCallsFormat}
\title{Assert KIR counts data frame format}
\usage{
checkKirCallsFormat(kir_calls)
}
\arguments{
\item{kir_calls}{KIR calls data frame, as returned by
\code{\link{readKIRCalls}} function.}
}
\value{
Logical indicating if \code{kir_calls} follow KIR counts data frame
  format. Otherwise raise an error.
}
\description{
\code{checkKirCallsFormat} asserts if KIR counts data frame have proper
format.
}
