% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_utils.R
\name{checkKirCountsFormat}
\alias{checkKirCountsFormat}
\title{Assert KIR counts data frame format}
\usage{
checkKirCountsFormat(kir_counts, accept.null = FALSE)
}
\arguments{
\item{kir_counts}{Data frame containing KIR gene counts, as returned by
\code{\link{readKirCalls}} function.}

\item{accept.null}{Logical indicating if NULL \code{kir_counts} should be
accepted.}
}
\value{
Logical indicating if \code{kir_counts} follow KIR counts data frame
  format. Otherwise raise error.
}
\description{
\code{checkKirCountsFormat} asserts if KIR counts data frame have proper
format.
}
\examples{
file <- system.file("extdata", "KIP_output_example.txt", package = "MiDAS")
kir_counts <- readKirCalls(file)
checkKirCountsFormat(kir_counts)

}
\seealso{
\code{\link{readKirCalls}}, \code{\link{getHlaKirInteractions}},
  \code{\link{kirHaplotypeToCounts}}, \code{\link{prepareMiDAS}}.

Other assert functions: \code{\link{characterMatches}},
  \code{\link{checkAdditionalData}},
  \code{\link{checkColDataFormat}},
  \code{\link{checkHlaCallsFormat}},
  \code{\link{checkKirCallsFormat}},
  \code{\link{checkStatisticalModel}},
  \code{\link{colnamesMatches}},
  \code{\link{functionExists}},
  \code{\link{hasTidyMethod}},
  \code{\link{isCharacterOrNULL}},
  \code{\link{isClassOrNULL}}, \code{\link{isClass}},
  \code{\link{isCountOrNULL}},
  \code{\link{isCountsOrZeros}},
  \code{\link{isExperimentCountsOrZeros}},
  \code{\link{isFlagOrNULL}}, \code{\link{isNumberOrNULL}},
  \code{\link{isStringOrNULL}},
  \code{\link{isTRUEorFALSE}},
  \code{\link{objectHasPlaceholder}},
  \code{\link{stringMatches}},
  \code{\link{validateFrequencyCutoffs}}
}
\concept{assert functions}
