% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_asserts.R
\name{checkStatisticalModel}
\alias{checkStatisticalModel}
\title{Assert statistical model}
\usage{
checkStatisticalModel(object)
}
\arguments{
\item{object}{An existing fit from a model function such as lm, glm and many
others.}
}
\value{
Logical indicating if \code{object} is an existing fit from a
model functions such as lm, glm and many others. Containing MiDAS object as
its data attribute. Otherwise raise an error.
}
\description{
\code{checkStatisticalModel} asserts if object is an existing fit from a
model functions such as lm, glm and many others. Containing MiDAS object as
its data atribute.
}
