% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_utils.R
\name{checkStatisticalModel}
\alias{checkStatisticalModel}
\title{Assert statistical model}
\usage{
checkStatisticalModel(object)
}
\arguments{
\item{object}{An existing fit from a model function such as lm, glm and many
others.}
}
\value{
Logical indicating if \code{object} is an existing fit from a
model function such as lm, glm and many others. ontaining MiDAS object as
its data atribute. Otherwise raise error.
}
\description{
\code{checkStatisticalModel} asserts if object is an existing fit from a
model function such as lm, glm and many others. Containing MiDAS object as
its data atribute.
}
\seealso{
Other assert functions: \code{\link{characterMatches}},
  \code{\link{checkAdditionalData}},
  \code{\link{checkColDataFormat}},
  \code{\link{checkHlaCallsFormat}},
  \code{\link{checkKirCallsFormat}},
  \code{\link{checkKirCountsFormat}},
  \code{\link{colnamesMatches}},
  \code{\link{functionExists}},
  \code{\link{hasTidyMethod}},
  \code{\link{isCharacterOrNULL}},
  \code{\link{isClassOrNULL}}, \code{\link{isClass}},
  \code{\link{isCountOrNULL}},
  \code{\link{isCountsOrZeros}},
  \code{\link{isFlagOrNULL}}, \code{\link{isNumberOrNULL}},
  \code{\link{isStringOrNULL}},
  \code{\link{isTRUEorFALSE}},
  \code{\link{objectHasPlaceholder}},
  \code{\link{stringMatches}}
}
\concept{assert functions}
