% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_asserts.R
\name{colnamesMatches}
\alias{colnamesMatches}
\title{Check column names}
\usage{
colnamesMatches(x, cols)
}
\arguments{
\item{x}{Data frame to test.}

\item{cols}{Ordered character vector to test against \code{x}'s colnames.}
}
\value{
Logical indicating if \code{x}'s colnames equals \code{choice}.
}
\description{
\code{colnamesMatches} check if  data frame's columns are named as specified
}
