% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_utils.R
\name{colnamesMatches}
\alias{colnamesMatches}
\title{Check column names}
\usage{
colnamesMatches(x, cols)
}
\arguments{
\item{x}{Data frame to test.}

\item{cols}{Ordered character vector to test against \code{x}'s colnames.}
}
\value{
Logical indicating if \code{x}'s colnames equals \code{choice}.
}
\description{
\code{colnamesMatches} check if  data frame's columns are named as specified
}
\seealso{
Other assert functions: \code{\link{characterMatches}},
  \code{\link{checkAdditionalData}},
  \code{\link{checkColDataFormat}},
  \code{\link{checkHlaCallsFormat}},
  \code{\link{checkKirCallsFormat}},
  \code{\link{checkKirCountsFormat}},
  \code{\link{checkStatisticalModel}},
  \code{\link{functionExists}},
  \code{\link{hasTidyMethod}},
  \code{\link{isCharacterOrNULL}},
  \code{\link{isClassOrNULL}}, \code{\link{isClass}},
  \code{\link{isCountOrNULL}},
  \code{\link{isCountsOrZeros}},
  \code{\link{isFlagOrNULL}}, \code{\link{isNumberOrNULL}},
  \code{\link{isStringOrNULL}},
  \code{\link{isTRUEorFALSE}},
  \code{\link{objectHasPlaceholder}},
  \code{\link{stringMatches}}
}
\concept{assert functions}
