% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_utils.R
\name{convertAlleleToVariable}
\alias{convertAlleleToVariable}
\title{Converts allele numbers to additional variables based on match file.}
\usage{
convertAlleleToVariable(allele, dictionary)
}
\arguments{
\item{allele}{Character vector containing HLA allele numbers.}

\item{path}{Path to the file containing HLA allele numbers matchings. The file
should be in tsv format with header and two columns. First column should hold
allele numbers and second corresponding additional variables. Function also
accepts match table in form of data frame.}
}
\value{
Vector containing allele numbers converted to additional variables
according to matching file.
}
\description{
Converts allele numbers to additional variables based on match file.
}
\examples{
dictionary <- system.file("extdata", "Match_4digit_supertype.txt", package = "MiDAS")
convertAlleleToVariable(c("A*01:01", "A*02:01"), dictionary = dictionary)

}
