% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dict_dist_grantham}
\alias{dict_dist_grantham}
\title{Grantham distance}
\format{An object of class \code{integer} of length 400.}
\usage{
dict_dist_grantham
}
\description{
Grantham distance
}
\keyword{datasets}
