% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drougth_classification.R
\name{drougth_classification}
\alias{drougth_classification}
\title{calculation of drougth classification}
\usage{
drougth_classification(r = call_n_pfiles(), p_files = call_Perc_files(),
  fun_ = drought_class)
}
\arguments{
\item{r}{a vector several raster directories (output of \link[MIDAS]{call_n_pfiles})}

\item{p_files}{percentile files (output of \link[MIDAS]{call_Perc_files})}

\item{fun_}{a function that classify drougth}
}
\value{
a raster file object
}
\description{
calculation of drougth classification
}
\author{
Adrian Huerta
}
