% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_transformationFunctions.R
\name{filterExperimentByFrequency}
\alias{filterExperimentByFrequency}
\title{Filter experiment by frequency}
\usage{
filterExperimentByFrequency(experiment, inheritance_model = c("dominant",
  "recessive", "additive"), lower_frequency_cutoff = NULL,
  upper_frequency_cutoff = NULL)
}
\arguments{
\item{experiment}{Matrix of type integer}

\item{inheritance_model}{String matching one of values
\code{"dominant", "recessive", "additive"}}

\item{lower_frequency_cutoff}{Number of positive value or \code{NULL}}

\item{upper_frequency_cutoff}{Number of positive value or \code{NULL}}
}
\value{
Matrix
}
\description{
Helper function for experiments filtering
}
