% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_transformationFunctions.R
\name{filterExperimentByVariables}
\alias{filterExperimentByVariables}
\alias{filterExperimentByVariables.matrix}
\alias{filterExperimentByVariables.SummarizedExperiment}
\title{Filter experiment by variable}
\usage{
filterExperimentByVariables(experiment, variables)

\method{filterExperimentByVariables}{matrix}(experiment, variables)

\method{filterExperimentByVariables}{SummarizedExperiment}(experiment, variables)
}
\arguments{
\item{experiment}{Matrix or SummarizedExperiment object.}

\item{variables}{Character vector specifying features to choose.}
}
\value{
Filtered \code{experiment} object.
}
\description{
Helper function for experiments filtering
}
