% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_transformationFunctions.R
\name{formatAssociationsResults}
\alias{formatAssociationsResults}
\title{Pretty format association analysis results}
\usage{
formatAssociationsResults(results, type = "hla_allele",
  response_variable = "R", logistic = FALSE, pvalue_cutoff = NULL,
  format = getOption("knitr.table.format"))
}
\arguments{
\item{results}{Tibble as returned by \link{analyzeAssociations}.}

\item{type}{String specifying type of analysis from which \code{results} were
produced. Possible values includes \code{'hla_allele'}, \code{'aa_level'},
\code{'expression_level'}.}

\item{response_variable}{String giving the name of response variable, it is
used to produce binary phenotype column names.}

\item{logistic}{Logical indicating if statistical model is logistic. If set
to \code{TRUE}, estimate will be renamed to odds ratio.}

\item{pvalue_cutoff}{Number specifying p-value cutoff for results to be
included in output. If \code{NULL} cutoff of \code{0.05} on
\code{p.adjusted} value is used instead.}

\item{format}{A character string. Possible values are \code{"latex"} and
\code{"html"}.}
}
\value{
A character vector with pretty formatted \code{results} table.
}
\description{
\link{formatAssociationsResults} formats results table to specified
format. It uses \link{formatResults} with prespecifed arguments to return
nice table depending on the type of analysis and model type. This function is
intended only to be used internally by \link{analyzeMiDASData}.
}
