% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_transformationFunctions.R
\name{formatResults}
\alias{formatResults}
\title{Helper function for pretty formating statistical analysis results}
\usage{
formatResults(results, filter_by = "p.value <= 0.05",
  arrange_by = "p.value", select_cols = c("term", "estimate",
  "std.error", "p.value", "p.adjusted"), format = c("html", "latex"),
  header = NULL)
}
\arguments{
\item{results}{Tibble as returned by \link{analyzeAssociations}.}

\item{filter_by}{Character specifying conditional expression used to filter
\code{results}, this is equivalent to \code{...} argument passed to
\link[dplyr]{filter} except it has to be a character vector.}

\item{arrange_by}{Character specifying variable names to use for sorting.
Equivalent to \code{...} argument passed to \link[dplyr]{arrange}.}

\item{select_cols}{Character specifying variable names that should be
included in the output table. Can be also used to rename selected
variables, see examples.}

\item{format}{A character string. Possible values are \code{"latex"} and
\code{"html"}.}

\item{header}{String specifying header for result table. If NULL header is
omitted.}
}
\value{
A character vector of the table source code.
}
\description{
\code{formatResults} formats statistical analysis results table to html or
latex format.
}
\examples{
hla_calls <- readHlaCalls(system.file("extdata", "HLAHD_output_example.txt", package = "MiDAS"))
hla_counts <- hlaCallsToCounts(hla_calls, inheritance_model = "additive")
midas_data <- read.table(
  system.file("extdata", "pheno_example.txt", package = "MiDAS"),
  header = TRUE)
midas_data <- dplyr::left_join(x = midas_data, y = hla_counts, by = "ID")
object <- lm(OS ~ 1, data = midas_data)
res <- analyzeAssociations(object)
formatResults(res,
              filter_by = c("p.value <= 0.05", "estimate > 0"),
              arrange_by = c("p.value * estimate"),
              select_cols = c("allele" = "term", "p.value"),
              format = "html",
              header = "HLA allelic associations")

}
