% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_utils.R
\name{getAlleleResolution}
\alias{getAlleleResolution}
\title{Infers HLA allele resolution}
\usage{
getAlleleResolution(allele)
}
\arguments{
\item{allele}{Character vector containing HLA allele numbers.}
}
\value{
Integer vector specifiying alleles resolutions.
}
\description{
Infers HLA allele resolution
}
\examples{
allele <- c("A*01:01", "A*01:02")
getAlleleResolution(allele)

}
