% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_transformationFunctions.R
\name{getCountsFrequencies}
\alias{getCountsFrequencies}
\title{Calculate variables frequencies}
\usage{
getCountsFrequencies(counts_table)
}
\arguments{
\item{counts_table}{Data frame containing variables counts, such as produced
by \link{hlaCallsToCounts}.}
}
\value{
Data frame containing variables, its corresponding total counts
  and frequencies.
}
\description{
\code{getCountsFrequencies} calculate variables frequencies based on counts
table, such as produced by \link{hlaCallsToCounts}.
}
\details{
Variables frequencies are counted in reference to sample size, depending on
the inheritance model under which the counts table has been generated one
might need to take under consideration both gene copies. Here sample size is
assumed to be depended on both gene copies if any count is greater than
\code{1} (`n / (2 * j)` where `n` is the number of term occurrences and `j`
is the sample size). If this is not the case the sample size is taken as is
(`n / j`).
}
\examples{
file <- system.file("extdata", "HLAHD_output_example.txt", package = "MiDAS")
hla_calls <- readHlaCalls(file)
hla_counts <- hlaCallsToCounts(hla_calls, inheritance_model = "additive")
getCountsFrequencies(hla_counts)

}
