% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_transformationFunctions.R
\name{getExperimentFrequencies}
\alias{getExperimentFrequencies}
\alias{getExperimentFrequencies.matrix}
\alias{getExperimentFrequencies.SummarizedExperiment}
\title{Calculate experiment's features frequencies}
\usage{
getExperimentFrequencies(experiment, carrier_frequency = FALSE,
  ref = NULL)

\method{getExperimentFrequencies}{matrix}(experiment,
  carrier_frequency = FALSE, ref = NULL)

\method{getExperimentFrequencies}{SummarizedExperiment}(experiment,
  carrier_frequency = TRUE, ref = NULL)
}
\arguments{
\item{experiment}{Matrix or SummarizedExperiment object.}

\item{carrier_frequency}{Logical flag indicating if carrier frequency should
be returned.}

\item{ref}{Wide format data frame with first column named "var" holding
features matching \code{experiment} and specific populations frequencies in
following columns. See \code{\link{getReferenceFrequencies}} for more
details.}
}
\value{
Data frame containing variables and their corresponding frequencies.
}
\description{
\code{getExperimentFrequencies} calculate features frequencies.
}
