% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_transformationFunctions.R
\name{getExperimentFrequencies}
\alias{getExperimentFrequencies}
\title{Calculate variables frequencies TODO}
\usage{
getExperimentFrequencies(experiment, inheritance_model = c("dominant",
  "recessive", "additive"))
}
\arguments{
\item{experiment}{Matrix}

\item{inheritance_model}{String}
}
\value{
Data frame containing variables, its corresponding total counts
  and frequencies.
}
\description{
\code{getExperimentFrequencies} calculate variables frequencies based on counts
table, such as produced by \code{\link{hlaCallsToCounts}}.
}
\details{
Variables frequencies are counted in reference to sample size, depending on
the inheritance model under which the counts table has been generated one
might need to take under consideration both gene copies. Here sample size is
assumed to be depended on both gene copies for \code{"additive"} inheritance
model (`n / (2 * j)` where `n` is the number of term occurrences and `j`
is the sample size). For other models the sample size is taken as is
(`n / j`).
}
\examples{
\dontrun{
file <- system.file("extdata", "HLAHD_output_example.txt", package = "MiDAS")
hla_calls <- readHlaCalls(file)
hla_counts <- hlaCallsToCounts(hla_calls, inheritance_model = "additive")
getExperimentFrequencies(hla_counts)
}

}
\seealso{
\code{\link{hlaCallsToCounts}}
}
