% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_transformationFunctions.R
\name{getFrequencyMask}
\alias{getFrequencyMask}
\title{Helper function for frequency data frame filtering}
\usage{
getFrequencyMask(df, lower_frequency_cutoff = NULL,
  upper_frequency_cutoff = NULL)
}
\arguments{
\item{df}{Data frame as returned by \code{getExperimentFrequencies}.}

\item{lower_frequency_cutoff}{Number of positive value or \code{NULL}}

\item{upper_frequency_cutoff}{Number of positive value or \code{NULL}}
}
\value{
Character vector containing names of terms after filtration.
}
\description{
Helper function for frequency data frame filtering
}
