% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_class.R
\docType{methods}
\name{getHlaCalls}
\alias{getHlaCalls}
\alias{getHlaCalls,MiDAS-method}
\title{Extract HLA calls from MiDAS object.}
\usage{
getHlaCalls(object)

\S4method{getHlaCalls}{MiDAS}(object)
}
\arguments{
\item{object}{\code{\link{MiDAS}} object}
}
\value{
Data frame with object's HLA calls.
}
\description{
Extract HLA calls from MiDAS object.
}
