% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_transformationFunctions.R
\name{getHlaFrequencies}
\alias{getHlaFrequencies}
\title{Calculate alleles frequencies}
\usage{
getHlaFrequencies(hla_calls, carrier_frequency = FALSE,
  compare = FALSE, ref_pop = c("USA NMDP African American pop 2",
  "USA NMDP Chinese", "USA NMDP European Caucasian"),
  ref = allele_frequencies)
}
\arguments{
\item{hla_calls}{Data frame containing HLA allele calls, as return by
\code{\link{readHlaCalls}} function.}
}
\value{
Data frame containing alleles and thier corresponding frequencies.
}
\description{
\code{getHlaFrequencies} calculates alleles frequencies in HLA calls data
frame.
}
\details{
Allele frequencies are counted in reference to sample taking both gene copies
into consideration. `n / (2 * j)` where `n` is the number of allele
occurrences and `j` is the sample size.
}
\examples{
file <- system.file("extdata", "HLAHD_output_example.txt", package = "MiDAS")
hla_calls <- readHlaCalls(file)
getHlaFrequencies(hla_calls)

}
