% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_transformationFunctions.R
\name{getHlaKirInteractions}
\alias{getHlaKirInteractions}
\title{Get HLA - KIR interactions}
\usage{
getHlaKirInteractions(hla_calls, kir_counts,
  interactions_dict = system.file("extdata", "Interactions_hla_kir.tsv",
  package = "MiDAS"))
}
\arguments{
\item{hla_calls}{Data frame containing HLA allele calls, as return by
\code{\link{readHlaCalls}} function.}

\item{kir_counts}{Data frame containing KIR genes counts, as return by
\link{readKirCalls}.}

\item{interactions_dict}{Path to the file containing HLA - KIR interactions
matchings. See details for further details.}
}
\value{
Data frame with binary presence-absence indicators for HLA - KIR
  interactions.

In order to be able to compare input data with \code{interactions_dict}
\code{hla_calls} are first converted to variables such as G groups, using
matching files shipped with the packages. Moreover \code{hla_calls} are also
reduced to all possible resolutions.

\code{interactions_dict} should be a tsv formatted file with following
columns: HLA, KIR, Affinity, Type. The package is shipped with interactions
file created based on Pende et al., 2019.
}
\description{
\code{getHlaKirInteractions} calculates binary presence-absence matrix of HLA
- KIR interactions.
}
\examples{
hla_file <- system.file("extdata", "HLAHD_output_example.txt", package = "MiDAS")
hla_calls <- readHlaCalls(hla_file)
kir_file <- system.file("extdata", "KIP_output_example.txt", package = "MiDAS")
kir_counts <- readKirCalls(kir_file, counts = TRUE)
getHlaKirInteractions(hla_calls, kir_counts)

}
