% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_class.R
\docType{methods}
\name{getInheritanceModel}
\alias{getInheritanceModel}
\alias{getInheritanceModel,MiDAS-method}
\title{Extract inheritance model from MiDAS object.}
\usage{
getInheritanceModel(object)

\S4method{getInheritanceModel}{MiDAS}(object)
}
\arguments{
\item{object}{\code{\link{MiDAS}} object}
}
\value{
String name of object's inheritance model.
}
\description{
Extract inheritance model from MiDAS object.
}
