% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_class.R
\name{getPlaceholder}
\alias{getPlaceholder}
\title{Get placeholder name from MiDAS object.}
\usage{
getPlaceholder(object)
}
\arguments{
\item{object}{\code{\link{MiDAS}} object.}
}
\value{
String giving name of placeholder.
}
\description{
Get placeholder name from MiDAS object.
}
