% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_utils.R
\name{getVariableAAPos}
\alias{getVariableAAPos}
\title{Returns positions of variable aa in the alignment.}
\usage{
getVariableAAPos(alignment)
}
\arguments{
\item{alignment}{Matrix containing aa level alignment.}
}
\value{
Integer vector specifying which alignment columns are variable.
}
\description{
Returns positions of variable aa in the alignment.
}
\examples{
file <- system.file("extdata", "A_prot.txt", package = "MiDAS")
alignment <- readHlaAlignments(file)
getVariableAAPos(alignment)

}
