% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_asserts.R
\name{hasTidyMethod}
\alias{hasTidyMethod}
\title{Check if tidy method for class exist}
\usage{
hasTidyMethod(class)
}
\arguments{
\item{class}{String giving object class.}
}
\value{
Logical indicating if there is a tidy method for a given class.
}
\description{
\code{hasTidyMethod} check if there is a tidy method available for a given class.
}
\seealso{
Other assert functions: \code{\link{checkColDataFormat}},
  \code{\link{isClassOrNULL}}, \code{\link{isCountOrNULL}},
  \code{\link{isExperimentCountsOrZeros}},
  \code{\link{isTRUEorFALSE}},
  \code{\link{objectHasPlaceholder}}
}
\concept{assert functions}
