% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_transformationFunctions.R
\name{hlaCallsToCounts}
\alias{hlaCallsToCounts}
\title{Transform HLA calls to counts table}
\usage{
hlaCallsToCounts(hla_calls)
}
\arguments{
\item{hla_calls}{Data frame containing HLA allele calls, as return by
\code{\link{readHlaCalls}} function.}
}
\value{
Data frame containing counts of HLA alleles found in input data
  frame.
}
\description{
\code{hlaCallsToCounts} converts HLA calls data frame into counts table
}
\examples{
file <- system.file("extdata", "HLAHD_output_example.txt", package = "MiDAS")
hla_calls <- readHlaCalls(file)
hlaCallsToCounts(hla_calls)

}
