% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_transformationFunctions.R
\name{hlaToAAVariation}
\alias{hlaToAAVariation}
\title{Converts HLA allele numbers to amino acid variation}
\usage{
hlaToAAVariation(hla_calls, indels = TRUE, unkchar = FALSE,
  alnpath = system.file("extdata", package = "MiDAS"), as_df = TRUE)
}
\arguments{
\item{hla_calls}{Data frame containing HLA allele calls, as return by
\code{\link{readHlaCalls}} function.}

\item{indels}{Logical indicating whether indels should be considered as
variability.}

\item{unkchar}{Logical indicating whether unknown characters in the alignment
should be treated as variability.}

\item{alnpath}{String providing optional path to directory containing HLA
alignment files. See details for further explanations.}

\item{as_df}{Logical indicating if data frame should be returned.
Otherwise function matrix is returned.}
}
\value{
Data frame or matrix containing variable amino acid positions. See
  \code{as_df} parameter.

  Rownames corresponds to ID column of input data frame, and colnames to
  alignment positions for given genes. If no variation in amino acids
  alignments is found function return one column matrix filled with `NA`.
}
\description{
\code{hlaToAAVariation} converts HLA allele numbers data frame to a matrix
holding information on amino acid level variation.
}
\details{
\code{alnpath} can be used to provide path to directory containing custom
alignment files. Each alignment file have to be named following EBI database
convention GENENAME_prot.txt. By default \code{alnpath} points to directory
containing alignments files shipped with the package.
}
\examples{
file <- system.file("extdata", "HLAHD_output_example.txt", package = "MiDAS")
hla_calls <- readHlaCalls(file)
hlaToAAVariation(hla_calls)

}
