% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_transformationFunctions.R
\name{hlaToAAVariation}
\alias{hlaToAAVariation}
\title{Converts HLA allele numbers to amino acid variation}
\usage{
hlaToAAVariation(hla_calls)
}
\arguments{
\item{hla_calls}{Data frame containing HLA allele calls.}
}
\value{
Matrix containing all variable amino acid positions.
}
\description{
Converts HLA allele numbers data frame to a data frame holding information on
amino acid level varation.
}
