% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_utils.R
\name{isCountOrNULL}
\alias{isCountOrNULL}
\title{Check if object is count or NULL}
\usage{
isCountOrNULL(x)
}
\arguments{
\item{x}{object to test.}
}
\value{
Logical indicating if object is count or NULL
}
\description{
\code{isCountOrNULL} check if object is a count (a single positive integer)
or NULL.
}
\seealso{
Other assert functions: \code{\link{characterMatches}},
  \code{\link{checkAdditionalData}},
  \code{\link{checkColDataFormat}},
  \code{\link{checkHlaCallsFormat}},
  \code{\link{checkKirCallsFormat}},
  \code{\link{checkStatisticalModel}},
  \code{\link{colnamesMatches}},
  \code{\link{functionExists}},
  \code{\link{hasTidyMethod}},
  \code{\link{isCharacterOrNULL}},
  \code{\link{isClassOrNULL}}, \code{\link{isClass}},
  \code{\link{isCountsOrZeros}},
  \code{\link{isExperimentCountsOrZeros}},
  \code{\link{isFlagOrNULL}},
  \code{\link{isInheritanceModelApplicable}},
  \code{\link{isNumberOrNULL}},
  \code{\link{isStringOrNULL}},
  \code{\link{isTRUEorFALSE}},
  \code{\link{objectHasPlaceholder}},
  \code{\link{stringMatches}},
  \code{\link{validateFrequencyCutoffs}}
}
\concept{assert functions}
