% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_utils.R
\name{isCountsOrZeros}
\alias{isCountsOrZeros}
\title{Check if vector contains only counts or zeros}
\usage{
isCountsOrZeros(x, na.rm = TRUE)
}
\arguments{
\item{x}{Numeric vector or object that can be \code{unlist} to numeric
vector.}

\item{na.rm}{Logical indicating if \code{NA} values should be accepted.}
}
\value{
Logical indicating if provided vector contains only positive integers
  or zeros.
}
\description{
\code{isCountsOrZeros} checks if vector contains only positive integers or
zeros.
}
\seealso{
Other assert functions: \code{\link{characterMatches}},
  \code{\link{checkAdditionalData}},
  \code{\link{checkHlaCallsFormat}},
  \code{\link{checkKirCountsFormat}},
  \code{\link{checkStatisticalModel}},
  \code{\link{colnamesMatches}},
  \code{\link{hasTidyMethod}},
  \code{\link{isCharacterOrNULL}},
  \code{\link{isClassOrNULL}}, \code{\link{isCountOrNULL}},
  \code{\link{isFlagOrNULL}}, \code{\link{isNumberOrNULL}},
  \code{\link{isStringOrNULL}},
  \code{\link{isTRUEorFALSE}},
  \code{\link{objectHasPlaceholder}},
  \code{\link{stringMatches}}
}
\concept{assert functions}
