% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_asserts.R
\name{isInheritanceModelApplicable}
\alias{isInheritanceModelApplicable}
\title{Check if experiment is inheritance model applicable}
\usage{
isInheritanceModelApplicable(x)
}
\arguments{
\item{x}{String giving experiment type.}
}
\value{
Logical indicating if the inheritance model can be applied.
}
\description{
Helper function checking if for given experiment type inheritance model can
be applied (eg. "hla_alleles") or not (eg. "hla_kir_interactions")
}
