% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_utils.R
\name{isNumberOrNULL}
\alias{isNumberOrNULL}
\title{Check if object is number or NULL}
\usage{
isNumberOrNULL(x)
}
\arguments{
\item{x}{object to test.}
}
\value{
Logical indicating if object is number or NULL
}
\description{
\code{isNumberOrNULL} checks if object is number (a length one numeric
vector) or NULL.
}
