% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_utils.R
\name{isStringOrNULL}
\alias{isStringOrNULL}
\title{Check if object is string or NULL}
\usage{
isStringOrNULL(x)
}
\arguments{
\item{x}{object to test.}
}
\value{
Logical indicating if object is string or NULL
}
\description{
\code{isStringOrNULL} checks if object is string (a length one character
vector) or NULL.
}
\seealso{
Other assert functions: \code{\link{characterMatches}},
  \code{\link{checkAdditionalData}},
  \code{\link{checkHlaCallsFormat}},
  \code{\link{checkKirCountsFormat}},
  \code{\link{checkStatisticalModel}},
  \code{\link{colnamesMatches}},
  \code{\link{hasTidyMethod}},
  \code{\link{isCharacterOrNULL}},
  \code{\link{isClassOrNULL}}, \code{\link{isCountOrNULL}},
  \code{\link{isCountsOrZeros}},
  \code{\link{isFlagOrNULL}}, \code{\link{isNumberOrNULL}},
  \code{\link{isTRUEorFALSE}},
  \code{\link{objectHasPlaceholder}},
  \code{\link{stringMatches}}
}
\concept{assert functions}
