% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_utils.R
\name{isTRUEorFALSE}
\alias{isTRUEorFALSE}
\title{Check if object is TRUE or FALSE flag}
\usage{
isTRUEorFALSE(x)
}
\arguments{
\item{x}{object to test.}
}
\value{
Logical indicating if object is TRUE or FALSE flag
}
\description{
\code{isTRUEorFALSE} checks if object is a flag (a length one logical vector)
except NA.
}
