% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_utils.R
\name{isTRUEorFALSE}
\alias{isTRUEorFALSE}
\title{Check if object is TRUE or FALSE flag}
\usage{
isTRUEorFALSE(x)
}
\arguments{
\item{x}{object to test.}
}
\value{
Logical indicating if object is TRUE or FALSE flag
}
\description{
\code{isTRUEorFALSE} check if object is a flag (a length one logical vector)
except NA.
}
\seealso{
Other assert functions: \code{\link{characterMatches}},
  \code{\link{checkAdditionalData}},
  \code{\link{checkColDataFormat}},
  \code{\link{checkHlaCallsFormat}},
  \code{\link{checkKirCallsFormat}},
  \code{\link{checkKirCountsFormat}},
  \code{\link{checkStatisticalModel}},
  \code{\link{colnamesMatches}},
  \code{\link{functionExists}},
  \code{\link{hasTidyMethod}},
  \code{\link{isCharacterOrNULL}},
  \code{\link{isClassOrNULL}}, \code{\link{isClass}},
  \code{\link{isCountOrNULL}},
  \code{\link{isCountsOrZeros}},
  \code{\link{isFlagOrNULL}}, \code{\link{isNumberOrNULL}},
  \code{\link{isStringOrNULL}},
  \code{\link{objectHasPlaceholder}},
  \code{\link{stringMatches}}
}
\concept{assert functions}
