% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_utils.R
\name{iterativeLRT}
\alias{iterativeLRT}
\title{Iterative likelihood ratio test}
\usage{
iterativeLRT(object, placeholder, omnibus_groups)
}
\arguments{
\item{object}{An existing fit from a model function such as lm, glm and many
others.}

\item{placeholder}{String specifying term to substitute with value from
\code{x}. Ignored if set to \code{NULL}.}

\item{omnibus_groups}{List of character vectors giving sets of variables for
which omnibus test should be applied.}
}
\value{
Data.frame containing summarised likelihood ratio test results.
}
\description{
\code{iterativeLRT} performs likelihood ratio test in an iterative manner
over groups of variables given in \code{omnibus_groups}.
}
