% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_utils.R
\name{kirHaplotypeToCounts}
\alias{kirHaplotypeToCounts}
\title{Convert KIR haplotypes to gene level}
\usage{
kirHaplotypeToCounts(x, hap_dict = system.file("extdata",
  "Match_KIR_haplotype_genes.tsv", package = "MiDAS"), binary = TRUE)
}
\arguments{
\item{x}{Character vector specifying KIR haplotypes.}

\item{hap_dict}{String specifying path to KIR haplotypes dictionary. By
default file shipped together with package is being used. See details for
more information.}

\item{binary}{Logical flag indicating if haplotypes should be converted only
  to gene presence / absence indicators. At this point this is the only way
  that allows unambiguous conversion.

\code{hap_dict} have to be a \code{tsv} file with first column holding KIR
haplotypes and gene counts in others. File should have header with first
column unnamed it is being used as row names.}
}
\value{
Data frame with haplotypes and corresponding gene counts. \code{NA}'s
  in \code{x} are removed during conversion.
}
\description{
\code{kirHaplotypeToCounts} converts vector of KIR haplotypes to data frame
of KIR genes counts.
}
\examples{
x <- c("1+3|16+3", "1+1", NA)
kirHaplotypeToCounts(x)

}
