% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_class.R
\name{MiDAS}
\alias{MiDAS}
\title{MiDAS class to hold all data and meta data about one MiDAS analysis.}
\usage{
MiDAS(hla_calls = NULL, kir_calls = NULL, colData,
  inheritance_model = c("additive", "dominant", "recesive"),
  analysis_type = c("hla_allele", "aa_level", "expression_level",
  "allele_g_group", "allele_supertype", "allele_group", "kir_genes",
  "hla_kir_interactions", "custom"), ...)

MiDAS(hla_calls = NULL, kir_calls = NULL, colData,
  inheritance_model = c("additive", "dominant", "recesive"),
  analysis_type = c("hla_allele", "aa_level", "expression_level",
  "allele_g_group", "allele_supertype", "allele_group", "kir_genes",
  "hla_kir_interactions", "custom"), ...)
}
\description{
The \link{\code{MiDAS}} class is a \link{\code{MultiAssayExperiment}} object
containing all the data and metadata about a set of measurments and thier
transformations required for MiDAS analysis.
}
\section{Slots}{

\describe{
\item{\code{metadata}}{A list that must at least contain a \code{inheritance_model}
  element.

The MiDAS object must contain at least \code{hla_calls} (see
\link{\code{readHlaCalls}}) or \code{kir_calls} (see
\link{\code{readKirCalls}}) experiments and defined \code{inheritance_model}
element of \code{metadata}.}
}}

\examples{
# read hla calls file
hla_calls_file <- system.file("extdata", "HLAHD_output_example.txt", package = "MiDAS")
hla_calls <- readHlaCalls(hla_calls_file)

# read batch covariates
pheno_file <- system.file("extdata", "pheno_example.txt", package = "MiDAS")
pheno <- read.table(pheno_file, header = TRUE, stringsAsFactors = FALSE)
covar_file <- system.file("extdata", "covar_example.txt", package = "MiDAS")
covar <- read.table(covar_file, header = TRUE, stringsAsFactors = FALSE)
colData <- left_join(pheno, covar, by ="ID")
rownames(coldata) <- coldata$ID

# create MiDAS object
midas <- MiDAS(hla_calls = hla_calls,
               colData = colData,
               inheritance_model = "additive"
)



}
