% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{midas_barplot}
\alias{midas_barplot}
\title{MIDAS Barplot}
\usage{
midas_barplot(midas_data, facet_by = "protein", FDR_cutoff = 0.1)
}
\arguments{
\item{midas_data}{\code{\link{midas_mpis}} with any filters applied}

\item{facet_by}{Variable to facet by, either protein or metabolite}

\item{FDR_cutoff}{cutoff used for labelling discoveries}
}
\value{
a grob
}
\description{
MIDAS Barplot
}
\examples{
midas_barplot(dplyr::filter(midas_mpis, query_protein \%in\% c("ACC", "AASSA-c213"), q_value < 0.1))

}
