% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_class.R
\name{midasToWide}
\alias{midasToWide}
\title{Transform MiDAS to wide format data.frame}
\usage{
midasToWide(object, experiment)
}
\arguments{
\item{object}{Object of class MiDAS}

\item{experiment}{Character}

\item{placeholder}{String}
}
\value{
Data frame
}
\description{
Transform MiDAS to wide format data.frame
}
