% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_asserts.R
\name{objectHasPlaceholder}
\alias{objectHasPlaceholder}
\title{Check if placeholder is present in object formula}
\usage{
objectHasPlaceholder(object, placeholder)
}
\arguments{
\item{object}{statistical model to test.}

\item{placeholder}{string specifying name of placeholder.}
}
\value{
Logical indicating if placeholder is present in object formula.
}
\description{
\code{isTRUEorFALSE} check if object is a flag (a length one logical vector)
except NA.
}
