% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_stats.R
\name{omnibusTest}
\alias{omnibusTest}
\title{Omnibus test}
\usage{
omnibusTest(object, omnibus_groups, placeholder = "term",
  correction = "bonferroni", n_correction = NULL)
}
\arguments{
\item{object}{An existing fit from a model function such as lm, glm and many
others.}

\item{omnibus_groups}{List of character vectors}

\item{placeholder}{String specyfing term in \code{object}'s formula which
should be substituted with an allele during analysis.}

\item{correction}{String specifying multiple testing correction method. See
details for further information.}

\item{n_correction}{Integer specifying number of comparisons to consider
during multiple testing correction calculations, must be at least equal to
number of comparisons being made; only set this (to non-default) when you
know what you are doing!}
}
\value{
Data frame containing omnibus test results for specified variables.
}
\description{
\code{OmnibusTest} calculate overall p-value for linear combination of
vairables using likelihood rato test.
}
\details{
Likelihood ratio test is conducted by comparing model given in \code{object}
with extended model, that is created by including effect of variables given
in \code{variables} as thie linear combination.
}
