% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_stats.R
\name{prepareHlaData}
\alias{prepareHlaData}
\title{Prepare data for statistical analysis}
\usage{
prepareHlaData(hla_calls, ..., inheritance_model = "additive")
}
\arguments{
\item{hla_calls}{Data frame containing HLA allele calls, as return by
\code{\link{readHlaCalls}} function.}

\item{...}{Data frames holding additional variables like phenotypic
  observations or covariates.

\code{...} should be data frames with first column holding samples IDs and
named \code{ID}. Those should correspond to \code{ID} column in
\code{hla_calls}.}

\item{inheritance_model}{String specifying inheritance model to use.
Available choices are \code{"dominant"}, \code{"recessive"},
\code{"additive"}. In \code{"dominant"} model homozygotes and heterozygotes
are coded as \code{1}. In \code{"recessive"} model homozygotes are coded as
\code{1} and all other as \code{0}. In \code{"additive"} model homozygotes
are coded as \code{2} and heterozygotes as \code{1}.}
}
\value{
Data frame with hla counts and additional variables.
}
\description{
\code{prepareHlaData} binds HLA alleles calls data frame with additional data
frames like phenotypic observations or covariates, creating an input data for
further statistical analysis.
}
\examples{
hla_calls_file <- system.file("extdata", "HLAHD_output_example.txt", package = "MiDAS")
hla_calls <- readHlaCalls(hla_calls_file)
pheno_file <- system.file("extdata", "pheno_example.txt", package = "MiDAS")
pheno <- read.table(pheno_file, header = TRUE)
covar_file <- system.file("extdata", "covar_example.txt", package = "MiDAS")
covar <- read.table(covar_file, header = TRUE)
prepareHlaData(hla_calls = hla_calls,
               pheno = pheno,
               covar = covar,
               inheritance_model = "additive"
)

}
