% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_stats.R
\name{prepareHlaData}
\alias{prepareHlaData}
\title{Prepare data for statistical analysis}
\usage{
prepareHlaData(hla_calls, pheno, covar = NULL, zygo = FALSE,
  reduce_counts = FALSE)
}
\arguments{
\item{hla_calls}{Data frame containing HLA allele calls, as return by
\code{\link{readHlaCalls}} function.}

\item{pheno}{Data frame holding phenotypic response variables.}

\item{covar}{Data frame holding covariates.}

\item{zygo}{Flag indicating whether zygosity should be added to
covariates. See details for further explanations.}

\item{reduce_counts}{Flag indicating whether allele counts should be reduced
to presence / absence indicators. See details for further explanations.}
}
\description{
Prepare data for statistical analysis
}
\examples{
hla_calls_file <- system.file("extdata", "HLAHD_output_example.txt", package = "MiDAS")
hla_calls <- readHlaCalls(hla_calls_file)
pheno_file <- system.file("extdata", "pheno_example.txt", package = "MiDAS")
pheno <- read.table(pheno_file, header = TRUE)
covar_file <- system.file("extdata", "covar_example.txt", package = "MiDAS")
covar <- read.table(covar_file, header = TRUE)
prepareHlaData(hla_calls = hla_calls,
               pheno = pheno,
               covar = covar,
               zygo = FALSE,
               reduce_counts = FALSE
)

}
