% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_class.R
\name{prepareMiDAS}
\alias{prepareMiDAS}
\alias{prepareMiDAS_hla_allele}
\title{Prepare MiDAS object for statistical analysis}
\usage{
prepareMiDAS(hla_calls = NULL, kir_calls = NULL, phenotype,
  inheritance_model = c("additive", "dominant", "recessive"),
  analysis_type = c("hla_allele", "aa_level", "allele_g_group",
  "allele_supertype", "allele_group", "kir_genes", "hla_kir_interactions"),
  ...)

prepareMiDAS_hla_allele(hla_calls, inheritance_model, ...)
}
\arguments{
\item{hla_calls}{Data frame containing HLA allele calls, as return by
\code{\link{readHlaCalls}} function.}

\item{kir_calls}{Data frame containing KIR calls, as return by
\code{\link{readKirCalls}} function.}

\item{phenotype}{Data frame holding additional variables like phenotypic
observations or covariates. It have to contain \code{'ID'} column holding
samples indentifires corresponding to indentifires in \code{hla_calls} and
\code{kir_calls}. Importantly rows of \code{hla_calls} and
\code{kir_calls} without corresponding phenotype are discarded.}

\item{inheritance_model}{String specifying inheritance model to use.
Available choices are \code{"dominant"}, \code{"recessive"},
\code{"additive"}. In \code{"dominant"} model homozygotes and heterozygotes
are coded as \code{1}. In \code{"recessive"} model homozygotes are coded as
\code{1} and all other as \code{0}. In \code{"additive"} model homozygotes
are coded as \code{2} and heterozygotes as \code{1}.}

\item{analysis_type}{Character vector indicating analysis type for which data
should be prepared. Valid choices are \code{"hla_allele"},
\code{"aa_level"}, \code{"allele_g_group"}, \code{"allele_supertype"},
\code{"allele_group"}, \code{"kir_genes"}, \code{"hla_kir_interactions"}.
See details for further explanations.}

\item{...}{Attributes used in data transformation.}
}
\value{
Object of class \code{\link{MiDAS}}

Matrix
}
\description{
\code{prepareMiDAS} transform HLA alleles calls and KIR calls according
to selected analysis creating \code{\link{MiDAS}} object.
}
\details{
\code{'hla_allele'} - \code{hla_calls} are transformed into counts
under \code{inheritance_model} of choice (see \code{\link{hlaCallsToCounts}}
for  more details).
}
\examples{
# read hla calls file
hla_calls_file <- system.file("extdata", "HLAHD_output_example.txt", package = "MiDAS")
hla_calls <- readHlaCalls(hla_calls_file)

# read phenotypic data and covariates
pheno_file <- system.file("extdata", "pheno_example.txt", package = "MiDAS")
pheno <- read.table(pheno_file, header = TRUE, stringsAsFactors = FALSE)
covar_file <- system.file("extdata", "covar_example.txt", package = "MiDAS")
covar <- read.table(covar_file, header = TRUE, stringsAsFactors = FALSE)
phenotype <- left_join(pheno, covar, by ="ID")

# create MiDAS object
midas <- MiDAS(hla_calls = hla_calls,
               phenotype = phenotype,
               inheritance_model = "additive",
               analysis_type = "hla_allele"
)

}
