% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_class.R
\name{prepareMiDAS_aa_level}
\alias{prepareMiDAS_aa_level}
\title{Prepare MiDAS data on HLA amino acid level}
\usage{
prepareMiDAS_aa_level(hla_calls, inheritance_model, indels = TRUE,
  unkchar = FALSE, ...)
}
\arguments{
\item{hla_calls}{Data frame}

\item{inheritance_model}{String}

\item{...}{Not used}
}
\value{
Matrix
}
\description{
\code{hla_calls} are first converted to amino acid level, taking only
variable positions under consideration. Than variable amino acid positions
are transformed to counts under \code{inheritance_model} of choice (see
\code{\link{hlaToAAVariation}} and \code{\link{aaVariationToCounts}} for more
details).
}
