% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_class.R
\name{prepareMiDAS_allele_g_group}
\alias{prepareMiDAS_allele_g_group}
\title{Prepare MiDAS data on HLA allele's G groups level}
\usage{
prepareMiDAS_allele_g_group(hla_calls, inheritance_model, ...)
}
\arguments{
\item{hla_calls}{Data frame}

\item{inheritance_model}{String}

\item{...}{Not used}
}
\value{
Matrix
}
\description{
\code{hla_calls} are transformed to HLA alleles groups using G group
dictionary shipped with the package. Than they are transformed to counts and
\code{\link{hlaCallsToCounts}} for more details).
}
