% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_class.R
\name{prepareMiDAS_allele_supertype}
\alias{prepareMiDAS_allele_supertype}
\title{Prepare MiDAS data on HLA allele's supertypes level}
\usage{
prepareMiDAS_allele_supertype(hla_calls, inheritance_model, ...)
}
\arguments{
\item{hla_calls}{Data frame}

\item{inheritance_model}{String}

\item{...}{Not used}
}
\value{
Matrix
}
\description{
\code{hla_calls} are transformed to HLA alleles groups using supertypes
dictionary shipped with the package. Than they are transformed to counts
under \code{inheritance_model} of choice (see \code{\link{hlaToVariable}} and
\code{\link{hlaCallsToCounts}} for more details).
}
