% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_class.R
\name{prepareMiDAS_hla_divergence}
\alias{prepareMiDAS_hla_divergence}
\title{Prepare MiDAS data on HLA divergence level}
\usage{
prepareMiDAS_hla_divergence(hla_calls, hla_divergence_range, ...)
}
\arguments{
\item{hla_calls}{HLA calls data frame, as returned by
\code{\link{readHlaCalls}} function.}

\item{...}{Not used}
}
\value{
Matrix
}
\description{
Prepare MiDAS data on HLA divergence level
}
