% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_class.R
\name{prepareMiDAS_kir_genes}
\alias{prepareMiDAS_kir_genes}
\title{Prepare MiDAS data on KIR genes level}
\usage{
prepareMiDAS_kir_genes(kir_calls, ...)
}
\arguments{
\item{kir_calls}{Data frame}

\item{...}{Not used}
}
\value{
Matrix
}
\description{
\code{kir_counts} data frame is joined with other inputs.
}
