% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_parsingFunctions.R
\name{readHlaAlignments}
\alias{readHlaAlignments}
\title{Reads HLA allele alignments}
\usage{
readHlaAlignments(file)
}
\arguments{
\item{file}{Path to the file containing HLA allele alignments.}
}
\value{
Matrix containing HLA allele alignments.
}
\description{
Reads HLA allele alignments from file in msf format.
}
