% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_parsingFunctions.R
\name{readHlaAlignments}
\alias{readHlaAlignments}
\title{Reads HLA allele alignments}
\usage{
readHlaAlignments(file, gene = NULL, trim = TRUE, unkchar = "",
  resolution = 8)
}
\arguments{
\item{file}{Path to the file containing HLA allele alignments.}

\item{gene}{Character vector of length one specifying the name of a gene for
which alignment is required. All the protein alignment files from EBI
database are shipped with the package and this parameter can be used to
provide simpler access to those files. If it's set to \code{NULL} file
parameter is used instead.}

\item{trim}{Logical indicating if alignment should be trimmed to start codon
of the mature protein.}

\item{unkchar}{Character to be used to represent positions with unknown
sequence.}

\item{resolution}{Integer specifying the resoultion with which alignment
matrix should be returned.}
}
\value{
Matrix containing HLA allele alignments. Rownames corresponds to
allele numbers and columns to positions in the alignment. Sequences
following the termination codon are marked as empty character. Unknown
sequences are marked with a character of choice, that defaults to empty
character (""). Stop codons are represented by a hash (X). Insertion and
deletions are marked with period (.).
}
\description{
Reads HLA allele alignments from file in msf format.
}
\examples{
hla_alignments <- readHlaAlignments(gene = "A")

}
