% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_parsingFunctions.R
\name{readHlaCalls}
\alias{readHlaCalls}
\title{Reads data table with HLA allele calls}
\usage{
readHlaCalls(file)
}
\arguments{
\item{file}{Path to the file containing HLA allele calls.}
}
\value{
Data frame containing HLA allele calls.
}
\description{
Reads data table with HLA allele calls from file in tsv format.
}
\examples{
file <- system.file("extdata/HLAHD_output_example.txt", package = "MiDAS")
hla_calls <- readHlaCalls(file)

}
