% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_parsingFunctions.R
\name{readKPICalls}
\alias{readKPICalls}
\title{Reads data table with KIR haplotypes calls}
\usage{
readKPICalls(file, hap_dict = system.file("extdata",
  "Match_kir_haplotype_gene.txt", package = "MiDAS"), counts = TRUE,
  binary = TRUE, na.strings = c("", "NA"))
}
\arguments{
\item{file}{Path to input file.}

\item{hap_dict}{String specifying path to KIR haplotypes dictionary. By
default file shipped together with package is being used. See details for
more information.}

\item{counts}{Logical flag indicating if KIR haplotypes should be converted
to gene counts.}

\item{binary}{Logical flag indicating if haplotypes should be converted only
  to gene presence / absence indicators (it is the only way that allows
  unambiguous conversion). This argument is currently ignored.

\code{hap_dict} have to be a tab separated values formatted file with first
column holding KIR haplotypes and gene counts in others. File should have
header with first column unnamed and gene names in the others.}

\item{na.strings}{a character vector of strings which are to be
    interpreted as \code{\link{NA}} values.  Blank fields are also
    considered to be missing values in logical, integer, numeric and
    complex fields.  Note that the test happens \emph{after} 
    white space is stripped from the input, so \code{na.strings} 
    values may need their own white space stripped in advance.}
}
\value{
Data frame containing KIR haplotypes calls or corresponding gene
  counts.
}
\description{
\code{readKPICalls} reads table with KIR haplotypes calls from file.
}
\details{
Input file have to be a tsv formatted table with two columns and header.
First column should contain samples IDs, second column should hold
corresponding KIR haplotypes.
}
\examples{
file <- system.file("extdata", "KIP_output_example.txt", package = "MiDAS")
readKPICalls(file)

}
