% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_utils.R
\name{reduceAlleleResolution}
\alias{reduceAlleleResolution}
\title{Reduce HLA allele number}
\usage{
reduceAlleleResolution(allele, resolution = 4)
}
\arguments{
\item{allele}{Character vector containing HLA allele numbers.}

\item{resolution}{Numeric vector of length one specifying desired
resolution. Note if `resolution` is greater than resolution of any
\code{allele} elements, those elements will be returned unchanged.}
}
\value{
Character vector containing reduced HLA allele numbers, importantly
alleles containing optional suffixes are omitted and returned unchanged.
}
\description{
Reduce HLA allele number
}
\examples{
reduceAlleleResolution(c("A*01", "A*01:24", "C*05:24:55:54"), 2)

}
