% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_stats.R
\name{runMiDAS}
\alias{runMiDAS}
\alias{runMiDAS_linear}
\alias{runMiDAS_conditional}
\title{Run MiDAS statistical analysis}
\usage{
runMiDAS(object, experiment, conditional = FALSE, omnibus = FALSE,
  correction = "bonferroni", n_correction = NULL,
  exponentiate = FALSE, ...)

runMiDAS_linear(object, experiment, correction = "bonferroni",
  n_correction = NULL, exponentiate = FALSE, ...)

runMiDAS_conditional(object, experiment, correction = "bonferroni",
  n_correction = NULL, exponentiate = FALSE, th = 0.05,
  keep = FALSE, rss_th = 1e-07, ...)
}
\arguments{
\item{object}{An existing fit from a model function such as lm, glm and many
others.}

\item{experiment}{String indicating the experiment associated with
\code{object}'s \code{MiDAS} data to use. Valid values includes:
\code{"hla_allele"}, \code{"aa_level"}, \code{"allele_g_group"},
\code{"allele_supertype"}, \code{"allele_group"}, \code{"kir_genes"},
\code{"hla_kir_interactions"}. See \code{link{prepareMiDAS}} for more
informations.}

\item{conditional}{Logical flag,}

\item{omnibus}{Logical flag.}

\item{correction}{String specifying multiple testing correction method. See
details for further information.}

\item{n_correction}{Integer specifying number of comparisons to consider
during multiple testing correction calculations, must be at least equal to
number of comparisons being made; only set this (to non-default) when you
know what you are doing!}

\item{exponentiate}{Logical flag indicating whether or not to exponentiate
the coefficient estimates. Internally this is passed to
\code{\link[broom]{tidy}}. This is typical for logistic and multinomial
regressions, but a bad idea if there is no log or logit link. Defaults to
FALSE.}

\item{th}{Number specifying p-value threshold for a variable to be considered
significant.}

\item{keep}{Logical flag indicating if the output should be a list of results
resulting from each selection step. Default is to return only the final
result.}

\item{rss_th}{Number specifying residual sum of squares threshold at which
function should stop adding additional variables. As the residual sum of
squares approaches \code{0} the perfect fit is obtained making further
attempts at variables selection nonsense, thus function is stopped. This
behavior can be controlled using \code{rss_th}.}

\item{mode}{String indicating analysis mode. See details for further
explenations.}
}
\value{
Tibble containing analysis results.
}
\description{
\code{runMiDAS} perform association analysis on MiDAS data using mode and
statistical model specified by user. Function is intended for use with
\code{\link{prepareMiDAS}}. See examples section.
}
\details{
\code{experiment} is used to select an experiment from \code{MiDAS} object
associated with \code{object} using. In standard work flow data are first
processed using \code{\link{prepareMiDAS}} creating experiments with
transformed data for given \code{experiment}. Next user constructs the
statistical model using function of choice (eg. \code{lm}. \code{coxph}).
Than \code{runMiDAS} is used to evaluate specified model uder \code{mode}
of choice. See the details of different \code{mode}'s implmenetions for
more details.

\code{correction} specifies p-value adjustment method to use, common choice
is Benjamini & Hochberg (1995) (\code{"BH"}). Internally this is passed to
\link[stats]{p.adjust}. Check there to get more details.

statistical analysis is performed iteratively
  on each variable in selected experiment. This is done by substituting
  \code{placeholder} in the \code{object}'s formula with each variable in the
  experiment.

statistical analysis is performed in a
  stepwise conditional testing manner, adding the previous top-associated
  variable as a covariate to \code{object}'s formula. The analysis stops
  when there is no more siginifcant variabls, based on self-defined
  threshold. The p-values of variables are used as the selection criteria.
  This proces is reapeated for each variable available in the selected
  experiment. This is done by substituting \code{placeholder} in the
  \code{object}'s formula with each variable in the experiment.
}
\examples{
\dontrun{
# read hla calls file
hla_calls_file <- system.file("extdata", "HLAHD_output_example.txt", package = "MiDAS")
hla_calls <- readHlaCalls(hla_calls_file)

# read kir calls file
kir_calls_file <- system.file("extdata", "KIP_output_example.txt", package = "MiDAS")
kir_calls <- readKirCalls(kir_calls_file, counts = TRUE)

# read phenotypic data and covariates
pheno_file <- system.file("extdata", "pheno_example.txt", package = "MiDAS")
pheno <- read.table(pheno_file, header = TRUE, stringsAsFactors = FALSE)
covar_file <- system.file("extdata", "covar_example.txt", package = "MiDAS")
covar <- read.table(covar_file, header = TRUE, stringsAsFactors = FALSE)
phenotype <- left_join(pheno, covar, by ="ID")

# create MiDAS object
midas <- prepareMiDAS(hla_calls = hla_calls,
                      kir_calls = kir_calls,
                      colData = phenotype,
                      inheritance_model = "additive",
                      experiment = "hla_allele"
)

# constructs statistical model
object <- lm(OS ~ AGE + SEX + term, data = midas)

# run analysis
runMiDAS(object, mode = "linear", experiment = "hla_allele")
}

}
\seealso{
\code{\link[stats]{p.adjust}}, \code{\link[broom]{tidy}}

Other MiDAS statistical functions: \code{\link{analyzeAssociations}},
  \code{\link{analyzeConditionalAssociations}}
}
\concept{MiDAS statistical functions}
