% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_utils.R
\name{runMiDASGetVarsFreq}
\alias{runMiDASGetVarsFreq}
\title{Add variables frequencies to runMiDAS results}
\usage{
runMiDASGetVarsFreq(midas, experiment, test_covar)
}
\arguments{
\item{midas}{MiDAS object.}

\item{experiment}{String specifying experiment from \code{midas}.}

\item{test_covar}{String giving name of test covariate.}
}
\value{
Data frame
}
\description{
Helper adding variables frequencies to statistical results. Additionally for
binary test covariate frequencies per phenotype are added.
}
