% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_utils.R
\name{stringMatches}
\alias{stringMatches}
\title{Check if string matches one of possible values}
\usage{
stringMatches(x, choice)
}
\arguments{
\item{x}{string to test.}

\item{choice}{Character vector with possible values for \code{x}.}
}
\value{
Logical indicating if \code{x} matches one of the strings in
  \code{choice}.
}
\description{
\code{stringMatches} checks if string is equal to one of the choices.
}
