% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_summarise.R
\name{summariseAAPosition}
\alias{summariseAAPosition}
\title{Summarize amino acid position}
\usage{
summariseAAPosition(hla_calls, aa_pos, aln = NULL, na.rm = FALSE)
}
\arguments{
\item{hla_calls}{Data frame containing HLA allele calls, as return by
\code{\link{readHlaCalls}} function.}

\item{aa_pos}{String specifying gene and amio acid position, example
\code{"A_9"}.}

\item{aln}{Matrix containing amino acid sequence alignments as returned by
\code{\link{readHlaAlignments}} function. By default function will use
alignment files shipped with the package.}

\item{na.rm}{Logical flag indicating if \code{NA} values should be considered
for frequency calculations.}
}
\value{
Data frame containing HLA alleles, their corresponding amino acid
  residues and frequencies at requested position.
}
\description{
List HLA alleles and amino acid residues at a given position
}
\examples{
file <- system.file("extdata", "HLAHD_output_example.txt", package = "MiDAS")
hla_calls <- readHlaCalls(file)
summariseAAPosition(hla_calls, "A_9")

}
