% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_utils.R
\name{updateModel}
\alias{updateModel}
\title{Add new variables to statistical model}
\usage{
updateModel(object, x, placeholder = NULL, backquote = TRUE,
  collapse = " + ")
}
\arguments{
\item{object}{An existing fit from a model function such as lm, glm and many
others.}

\item{x}{Character vector specifying variables to be added to model.}

\item{placeholder}{String specifying term to substitute with value from
\code{x}. If more than one variable is passed in \code{x} all of them will
be bracketed. Ignored if set to \code{NULL}.}

\item{backquote}{Logical indicating if added variables should be quoted.
Elements of this vector are recycled over \code{x}. Only relevant if
\code{x} is of type character.}

\item{collapse}{String specifying how new characters should be added to
old formula. Only relevant if \code{placeholder} is specified.}
}
\value{
Updated fit of input model.
}
\description{
\code{updateModel} adds new variables to model and re-fit it.
}
\examples{
object <- lm(dist ~ 1, data = cars)
updateModel(object, "dist")

object <- lm(Sepal.Length ~ Petal.Width*Species, iris)
updateModel(object, c("Sepal.Width", "Petal.Length"), placeholde = "Petal.Width")

}
