% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_utils.R
\name{updateModel}
\alias{updateModel}
\title{Add new variables to statistical model}
\usage{
updateModel(object, x, backquote = TRUE, collapse = " + ")
}
\arguments{
\item{object}{An existing fit from a model function such as lm, glm and many
others.}

\item{x}{Character vector specifying variables to be added to model or a
formula giving a template which specifies how to update.}

\item{backquote}{Logical indicating if added variables should be quoted.
Longer than one element vectors are accepted as well, specifying which new
variables should be backquoted. Only relevant if x is of type character.}

\item{collapse}{Character specifying how new characters should be added to
old formula. Only relevant if x is of type character.}
}
\value{
Updated fit of input model.
}
\description{
\code{updateModel} will add new variables to model and re-fit it.
}
\examples{
object <- lm(dist ~ 1, data = cars)
updateModel(object, "dist")

}
