% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS_utils.R
\name{updateModel}
\alias{updateModel}
\title{Add new variables to statistical model}
\usage{
updateModel(object, x, backquote = TRUE, collapse = " + ")
}
\arguments{
\item{object}{An existing fit from a model function such as lm, glm and many
others.}

\item{x}{Character vector specifying variables to be added to model or a
formula giving a template which specifies how to update.}

\item{backquote}{Logical indicating if added variables should be quoted.
Longer than one element vectors are accepted as well, specifying which new
variables should be backquoted. Only relevant if x is of type character.}

\item{collapse}{Character specifying how new characters should be added to
old formula. Only relevant if x is of type character.}
}
\value{
Updated fit of input model.
}
\description{
\code{updateModel} will add new variables to model and re-fit it.
}
\examples{
library("survival")
hla_calls_file <- system.file("extdata", "HLAHD_output_example.txt", package = "MiDAS")
hla_calls <- readHlaCalls(hla_calls_file)
pheno_file <- system.file("extdata", "pheno_example.txt", package = "MiDAS")
pheno <- read.table(pheno_file, header = TRUE)
covar_file <- system.file("extdata", "covar_example.txt", package = "MiDAS")
covar <- read.table(covar_file, header = TRUE)
hla_data <- prepareHlaData(hla_calls, pheno, covar)
coxmod <- hlaAssocModel(model = "coxph",
                        response = "Surv(OS, OS_DIED)",
                        variable = "1",
                        data = hla_data$data
)
updateModel(coxmod, "A*01:01", backquote = TRUE, collapse = " + ")

}
